% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-class.R
\name{as.demography_dynamics}
\alias{as.demography_dynamics}
\title{Change the demography in a state object}
\usage{
as.demography_dynamics(demography_dynamics_function)
}
\arguments{
\item{demography_dynamics_function}{A function that operates on a state object to change demography at specified timesteps. User may enter a custom function or select a pre-defined module - see documentation.}
}
\value{
An object of class \code{demography_dynamics}
}
\description{
Change the demography in a state object
}
\examples{

library(raster)
library(dhmpr)

example_function <- function (state, timestep) {
state
}
no_demography_dynamics <- as.demography_dynamics(example_function)
}
