% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-class.R
\name{as.habitat_dynamics}
\alias{as.habitat_dynamics}
\title{Change the habitat in a state object}
\usage{
as.habitat_dynamics(habitat_dynamics_function)
}
\arguments{
\item{habitat_dynamics_function}{A function that operates on a state object to change habitat at specified timesteps. User may enter a custom function or select a pre-defined module - see documentation.}
}
\value{
An object of class \code{habitat_dynamics}
}
\description{
Change the habitat in a state object
}
\examples{

library(raster)
library(dhmpr)

example_function <- function (state, timestep) {
state
}
no_habitat_dynamics <- as.habitat_dynamics(example_function)
}
