% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{build_demography}
\alias{build_demography}
\title{Create a demography object to define population transition matrices}
\usage{
build_demography(transition_matrix, dispersal_parameters)
}
\arguments{
\item{transition_matrix}{A symmetrical stage-based population structure matrix}

\item{dispersal_parameters}{Specifications for dispersal in the landscape}
}
\value{
An object of class \code{demography}
}
\description{
Create a demography object to define population transition matrices
}
\examples{

library(raster)
library(dhmpr)

test_demography <- build_demography(fake_transition_matrix(4), rlnorm(1))
}
