% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics-class.R
\name{build_dynamics}
\alias{build_dynamics}
\title{Create a dynamics object to run in an experiment}
\usage{
build_dynamics(habitat_dynamics, demography_dynamics, population_dynamics,
  order = c("demography_dynamics", "habitat_dynamics", "population_dynamics"))
}
\arguments{
\item{habitat_dynamics}{A module to alter the habitat object in an experiment}

\item{population_dynamics}{A module to alter the population object in an experiment}

\item{demographic_dynamics}{A module to alter the habitat object in an experiment}
}
\value{
An object of class \code{dynamics}
}
\description{
Create a dynamics object to run in an experiment
}
\examples{

library(raster)
library(dhmpr)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
test_demography <- build_demography(fake_transition_matrix(4), rlnorm(1))
test_population <- build_population(stack(replicate(4, test_habitat$carrying_capacity * 0.2)))
test_state <- build_state(test_habitat, test_demography, test_population)
fast_approximation <- build_dynamics(no_habitat_dynamics, no_demographic_dynamics, fast_population_dynamics)
}
