% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{build_habitat}
\alias{build_habitat}
\title{Create a habitat object to contain spatial information on habitat suitability and carrying capacity of a landscape}
\usage{
build_habitat(habitat_suitability, carrying_capacity)
}
\arguments{
\item{habitat_suitability}{A raster layer or stack containing habitat suitability for each cell}

\item{carrying_capacity}{A raster layer specifying carrying capacity values for each cell}
}
\value{
An object of class \code{habitat}
}
\description{
Create a habitat object to contain spatial information on habitat suitability and carrying capacity of a landscape
}
\examples{

library(raster)
library(dhmpr)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
}
