% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{build_population}
\alias{build_population}
\title{Create a population object to use in a state object}
\usage{
build_population(population_raster)
}
\arguments{
\item{population_raster}{A raster stack with one layer for each life stage}
}
\value{
An object of class \code{population}
}
\description{
Create a population object to use in a state object
}
\examples{

library(raster)
library(dhmpr)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
test_population <- build_population(stack(replicate(4, test_habitat$carrying_capacity * 0.2)))
}
