% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-class.R
\name{build_state}
\alias{build_state}
\title{Create a state object to represent a static representation of habitat, population, and demography in a timestep}
\usage{
build_state(habitat, demography, population)
}
\arguments{
\item{habitat}{A habitat object - a habitat suitability raster layer or stack}

\item{demography}{A demography object - a stage-based transition matrix}

\item{population}{A population object - a raster stack with layers for each population stage}
}
\value{
An object of class \code{state}
}
\description{
Create a state object to represent a static representation of habitat, population, and demography in a timestep
}
\examples{

library(raster)
library(dhmpr)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
test_demography <- build_demography(fake_transition_matrix(4), rlnorm(1))
test_population <- build_population(stack(replicate(4, test_habitat$carrying_capacity * 0.2)))
test_state <- build_state(test_habitat, test_demography, test_population)
}
