% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_fun.R
\name{cc_fun}
\alias{cc_fun}
\title{calculate carrying capacity from occupancy}
\usage{
cc_fun(x, a = 6, b = 3, c = 0.5, cc_mod = c("exp", "logit", "linear"))
}
\arguments{
\item{x}{a raster of species habitit suitability (occupancy)}

\item{a}{parameter for carrying capacity model}

\item{b}{parameter for carrying capacity model}

\item{c}{parameter for carrying capacity model}

\item{cc_mod}{model form for converting hs to carrying capacity.}
}
\description{
calculate carrying capacity from occupancy
}
\author{
Skipton Woolley
}

