% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{count_rank_fun}
\alias{count_rank_fun}
\title{Initial ranking function}
\usage{
count_rank_fun(ind, ord.pairs)
}
\arguments{
\item{ind}{observation index to look for in \code{ord.pairs}}

\item{ord.pairs}{a table with sorted pairs of observations.}
}
\value{
The function counts for a specific observation how many times it was smaller than
         other observations and how many times it was greater.
}
\description{
Initial ranking function
}
\details{
As a good starting point for y values we will use the information we have on our sorted pairs.
         An observations will get a higher rank for starter if they appear a lot of times in a
        severe state comapring to others observations.
}
\examples{
\dontrun{
set.seed(6324)
pairs <- t(combn(x = c(1:10), m = 2))
ord.pairs <- t(apply(pairs, 1, ord_pairs_fun, sort.by = y))
count_rank_fun(ind = 1, ord.pairs)
}
}
