% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{demo_proj}
\alias{demo_proj}
\title{Single time step demographic projection function in C++}
\usage{
demo_proj(v0, tmat, matsd = NULL, stmat = NULL, estamb = FALSE,
  estdem = FALSE, equalsign = TRUE, tmat_fecundity = FALSE)
}
\arguments{
\item{v0}{NumericVector. Vector with the initial abundance of each stage.}

\item{tmat}{NumericMatrix. Transition matrix.}

\item{matsd}{NumericMatrix. Matrix with the standard deviation of the probabilities in tmat.}

\item{stmat}{NumericMatrix. Matrix indicating for each transition probability in mat which part (i.e. which proportion) should be considered resulting from fecundity.}

\item{estamb}{bool. Environmental stochasticity included in population dynamics?}

\item{estdem}{bool. Demographic stochasticity included in population dynamics?}

\item{equalsign}{bool. Should the environmental deviations have all the same sign and magnitude?}

\item{tmat_fecundity}{bool. Should the first row of tmat as fecundities? TRUE or FALSE}
}
\description{
Single time step demographic projection function in C++
}

