% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{demo_proj_n_cpp}
\alias{demo_proj_n_cpp}
\title{Multiple time step and repetition demographic projection function in C++}
\usage{
demo_proj_n_cpp(vn, tmat, matsd = NULL, stmat = NULL, estamb = FALSE,
  estdem = FALSE, equalsign = TRUE, tmat_fecundity = FALSE, nrep = 1L,
  time = 10L)
}
\arguments{
\item{vn}{List. List of vectors for simulation.}

\item{tmat}{NumericMatrix. Transition matrix.}

\item{matsd}{NumericMatrix. Transtion matrix error.}

\item{stmat}{NumericMatrix. Matrix indicating for each transition probability in mat which part (i.e. which proportion) should be considered resulting from fecundity}

\item{estamb}{bool. Environmental stochasticity included in population dynamics?}

\item{estdem}{bool. Demographic stochasticity included in population dynamics?}

\item{equalsign}{Logical. Should the environmental deviations have all the same sign and magnitude?}

\item{tmat_fecundity}{bool Should the first row of tmat as fecundities? TRUE or FALSE}

\item{nrep}{int number of simulations}

\item{time}{int number of time-steps.}
}
\description{
Multiple time step and repetition demographic projection function in C++
}

