% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic-class.R
\name{demographic}
\alias{demographic}
\alias{plot.demographic}
\title{demographic objects}
\usage{
demographic(v0, tmat, matsd = NULL, stmat = NULL, estamb = FALSE,
  estdem = FALSE, equalsign = TRUE, tmat_fecundity = FALSE, nrep = 10,
  time = 10)

\method{plot}{demographic}(object, mean_pop = TRUE, ...)
}
\arguments{
\item{v0}{vector Starting abundances at each timestep.}

\item{tmat}{matrix. Transition matrix}

\item{matsd}{matrix. Matrix with the standard deviation of the probabilities in tmat.}

\item{stmat}{martix. Matrix indicating for each transition probability in mat which part (i.e. which proportion) should be considered resulting from fecundity.}

\item{estamb}{Logical. Should environmental stochasticity be considered to projet the dynamics of the population?}

\item{estdem}{Logical. Should demographic stochasticity be employed to project the dynamics of the population?}

\item{equalsign}{Logical. Should the environmental deviations have all the same sign and magnitude?}

\item{tmat_fecundity}{Logical. Should you use the first row of tmat as fecundities?}

\item{nrep}{int number of simulations to run.}

\item{time}{int length of the demographic trajectory.}

\item{object}{a demographic_mod object}

\item{mean_pop}{logical. If TRUE plots the mean demographic change in population through time. Otherwise it plots the population changes for each stage in the transition matrix.}

\item{...}{other plot arguments}
}
\value{
demographic_mod object a list of demographic projections
}
\examples{
tmat <- as.transition_matrix(matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE))
matsd <- tmat/10
v0 <- c(80,20,0)
sim_t10_rep100 <- demographic(v0=v0,tmat=tmat,matsd=matsd,estdem=TRUE,time=10,nrep=100) 
tmat <- as.transition_matrix(matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE))
matsd <- tmat/10
v0 <- c(80,20,0)
sim_t20_rep100 <- demographic(v0=v0,tmat=tmat,matsd=matsd,estdem=TRUE,time=20,nrep=100) 
plot(sim_t20_rep100,mean_pop=TRUE)
plot(sim_t20_rep100,mean_pop=FALSE)
}
\author{
Skipton Woolley

Skipton Woolley
}
\seealso{
\code{link{demo_proj_n_cpp}}
}

