% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-class.R
\name{demography_dynamics}
\alias{demography_dynamics}
\alias{is.demography_dynamics}
\alias{print.demography_dynamics}
\title{Modify the demography in a state object.}
\usage{
demography_dynamics(...)

is.demography_dynamics(x)

\method{print}{demography_dynamics}(x, ...)
}
\arguments{
\item{...}{Functions that operates on a state object to change demography
at specified timesteps. A user may enter custom functions or select
pre-defined modules - see examples.}

\item{x}{A \code{demography_dynamics} object to print or test.}
}
\value{
An object of class \code{demography_dynamics}
}
\description{
A \code{demography_dynamics} object is used to modify life-stage transition
matrices - for example, adding stochasticity.
}
\details{
A \code{demography_dynamics} object is a sub-component of a \link[steps]{dynamics}
object and is executed in each timestep of a simulation. Note, some dynamics
functions can be executed at non-regular intervals (i.e. only timesteps
explicitly defined by the user). The \code{build_demography_dynamics} function is
used to construct a demography dynamics object consisting of several demographic
dynamics functions and their associated parameters. These functions specify how
the demography in the state object will be modified throughout a simulation.
}
\examples{

library(steps)
library(raster)

# Select existing dynamic functions to be run on the demography
# in a simulation and specify input parameters: 
env_stoch <- environmental_stochasticity(transition_matrix = egk_mat,
                                         stochasticity = 0.5)
                                              
dem_dens <- density_dependence(transition_matrix = egk_mat)

# Construct a demography dynamics object
dem_dynamics <- demography_dynamics(env_stoch, dem_dens)

# Test if object is of the type 'demography_dynamics'
is.demography_dynamics(test_demo_dynamics)

# Print details about the 'demography_dynamics' object
print(test_demo_dynamics)
}
