% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-functions.R
\name{demography_dynamics_functions}
\alias{demography_dynamics_functions}
\alias{demo_environmental_stochasticity}
\alias{demo_density_dependence}
\alias{demo_surv_fec_modify}
\title{Functions to modify the demography in a state object.}
\usage{
demo_environmental_stochasticity(transition_matrix, stochasticity = 0)

demo_density_dependence(transition_matrix, fecundity_fraction = 1,
  survival_fraction = 1)

demo_surv_fec_modify(transition_matrix, surv_layers, fec_layers)
}
\arguments{
\item{transition_matrix}{A life-stage transition matrix.}

\item{stochasticity}{A matrix with standard deviations (consistent or
varying) around the transition means with dimensions matched to the
life-stage transition matrix or a number representing a consistent
standard deviation to apply to all transitions (default is 0).}

\item{fecundity_fraction}{A multiplier value between 0 and 1 for fecundity
values in the transition matrix.}

\item{survival_fraction}{A multiplier value between 0 and 1 for survival
values in the transition matrix.}

\item{surv_layers}{a list of raster stacks with multipliers for survival
equal to the number of life-stages.}

\item{fec_layers}{a list of raster stacks with multipliers for fecundities
equal to the number of life-stages. Note, life-stages that do not reproduce
will have NULL in place of the raster stack.}
}
\description{
Pre-defined functions to operate on a population demography
during a simulation.
}
\examples{

library(steps)

# Use the demo_environmental_stochasticity function to modify the transition
# matrix with specified environmental stochasticity:
test_demo_es <- demo_environmental_stochasticity(transition_matrix = mat,
                                    stochasticity = mat_sd)

# Use the demo_density_dependence function to modify the transition
# matrix once carrying capacity is reached:
test_demo_dd <- demo_density_dependence(transition_matrix = mat,
                                        fecundity_fraction = 1,
                                        survival_fraction = 0.5)

# Use the demo_surv_fec_modify function to modify the  
# demography using explicit survival and fecundity layers:
test_survfec <- demo_surv_fec_modify(transition_matrix = mat,
                                    surv_layers = surv,
                                    fec_layers = fec)
}
