% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{deriv_fun}
\alias{deriv_fun}
\title{Derivatibe of Loss Objective Function according to F(x)}
\usage{
deriv_fun(Fx, n.B, setB, setW, setR, lambdaB, lambdaW, lambdaR, h)
}
\arguments{
\item{Fx}{Fx}

\item{n.B}{n.B}

\item{setB}{setB}

\item{setW}{setW}

\item{setR}{setR}

\item{lambdaB}{lambdaB}

\item{lambdaW}{lambdaW}

\item{lambdaR}{lambdaR}

\item{h}{h}
}
\value{
The function returns the value of the linear DSS model objective function.
}
\description{
Derivatibe of Loss Objective Function according to F(x)
}
\details{
The objective function is of the shape:
}
