% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal-class.R
\name{dispersal}
\alias{as.dispersal}
\alias{dispersal}
\alias{is.dispersal}
\title{dispersal class for metapoplation}
\usage{
dispersal(params)

is.dispersal(x)

as.dispersal(x)
}
\arguments{
\item{params}{list which containts alpha the exponential decay rate of patch connectivity 
(dispersion parameter), beta double parameter that represents the shape of the dispersal kernel.
and disp_fun a characture 'H' uses hanski(1994), if 'S' uses shaw(1995).}

\item{x}{an object to be tested as a dispersal transfun object}

\item{x}{an object to be tested as a dispersal transfun object}
}
\examples{
params <- list(alpha=1,beta=1,disp_fun="H")
dispersal(params)

dispersal(NULL)
is.dispersal(disp)
}

