% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel_functions-class.R
\name{dispersal_kernel}
\alias{dispersal_kernel}
\title{Create a dispersal function}
\description{
A dispersal kernel function is a mathematical representation of how species redistribute
across the landscape.
 
A common dispersal kernel is provided in the software for the user to select, however,
a user may also provide a custom written dispersal kernel.
}
\seealso{
\itemize{
  \item{\link[steps]{exponential_dispersal_kernel}) for a (negative) exponential dispersal
  kernel}
  }
}
