% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_proportion_functions-class.R
\name{all_dispersing}
\alias{all_dispersing}
\alias{carrying_capacity_dispersal}
\title{Create a proportion dispersing function}
\usage{
all_dispersing(proportions = 1)

carrying_capacity_dispersal()
}
\arguments{
\item{proportions}{proportions of individuals in each life stage that disperse - default is 1}
}
\value{
An object of class \code{dispersal_proportion_function}
}
\description{
A proportion dispersing function generates the proportions of species
that disperse from cells based on landscape features.
 
The default proportion dispersing function returns full dispersal (1.0) for all life stages.
Additional proportion dispersing functions are also provided in the software for the
user to select (see' \link[steps]{carrying_capacity_dispersal}), however, a user may also provide a
custom written proportion dispersing function.
}
\examples{

test_dispersal_function <- all_dispersing()

test_dispersal_function <- carrying_capacity_dispersal()
}
