% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{.ord_pairs_fun}
\alias{.ord_pairs_fun}
\title{Sorts pairs of observation to better (left) and severe (right) states}
\usage{
.ord_pairs_fun(pair, sort.by)
}
\arguments{
\item{pair}{a vector of length 2 with two observations indices.}

\item{sort.by}{A numeric or integer vector at the same length of number of observations to use for ordering. Higher value indiated severe state.}
}
\value{
The function returns the \code{pair} ordered from better (left) to severe (right) state according to \code{sort.by}
}
\description{
Sorts pairs of observation to better (left) and severe (right) states
}
\details{
The function checks which of the obsevations given in \code{pair} have the higher
}
\examples{
set.seed(263)
pairs <- t(combn(x = c(1:20), m = 2))
head(pairs)
# STEPS::ord_pairs_fun(pairs[1, ], sort.by = y)

}
