% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-class.R
\name{dynamic}
\alias{dynamic}
\alias{is.dynamic}
\title{dynamic object}
\usage{
dynamic(transition, population, ...)

is.dynamic(x)
}
\description{
Very similar to \link[pop]{dynamic}, \code{dynamic} is an object which stores all the relevent 
stage-matrices, transitions, population(s), habitat and dispersal for using the \link[dhmpr]{simulation}.
The main difference between \link[pop]{dynamic} and \code{dynamic} is that dhmpr requires the input of set objects to run.
}

