% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk}
\alias{egk}
\alias{egk_hab}
\alias{egk_pop}
\alias{egk_k}
\alias{egk_sf}
\alias{egk_fire}
\alias{egk_origins}
\alias{egk_destinations}
\alias{egk_road}
\alias{egk_mat}
\alias{egk_mat_stoch}
\title{Eastern Grey Kangaroo example data}
\format{
Misc data

An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
egk_hab()

egk_pop()

egk_k()

egk_sf()

egk_fire()

egk_origins()

egk_destinations()

egk_road()

egk_mat

egk_mat_stoch
}
\description{
Example data for simulating spatial population dynamics of Eastern Grey
Kangaroos in a hypothetical landscape.
}
\details{
\describe{
  \item{egk_hab}{A function that generates the SpatRaster containing the
   predicted relative habitat
  suitability for the Eastern Grey Kangaroo.}
  \item{egk_pop}{A function that generates the SpatRaster stack containing
   initial populations for each
  life-stage of the Eastern Grey Kangaroo.}
  \item{egk_k}{A function that generates the SpatRaster layer containing the
   total number of Eastern Grey
  Kangaroos each grid cell can support.}
  \item{egk_mat}{A matrix containing the survival and fecundity of Eastern
  Grey Kangaroos at each of three life-stages - juvenile, subadult, and 
  adult.}
  \item{egk_mat_stoch}{A matrix containing the uncertainty around survival
  and fecundity of Eastern Grey Kangaroos at each of three life-stages -
  juvenile, subadult, and adult.}
  \item{egk_sf}{A function that generates the SpatRaster stack containing 
  values for modifying survival and
  fecundities - each is raster is named according to the timestep and 
  position
  of the life-stage matrix to be modified.}
  \item{egk_fire}{A function that generates the SpatRaster stack containing 
  values for modifying the habitat
  - in this case the proportion of landscape remaining after fire.}
  \item{egk_origins}{A function that generates the SpatRaster stack 
  containing locations and counts of where
  to move individual kangaroos from.}
  \item{egk_destinations}{A function that generates the SpatRaster stack
   containing locations and counts of where to
  move individual kangaroos to.}
  \item{egk_road}{A function that generates the SpatRaster stack containing
   values for modifying the habitat
  - in this case the proportion of habitat remaining after the construction
   of a road.}
}
}
\keyword{datasets}
