% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_dist}
\alias{egk_dist}
\title{Eastern Grey Kangaroo habitat disturbance}
\format{A raster stack with 35 rows, 36 columns, 1260 cells, and 20 layers:
\describe{
  \item{resolution}{size of cells in meters}
  \item{extent}{cartesian boundaries in meters}
  \item{coord. ref.}{spatial projection of layer}
}}
\usage{
egk_dist
}
\description{
A raster stack containing values for modifying the habitat.
}
\keyword{datasets}
