% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_hab}
\alias{egk_hab}
\title{Eastern Grey Kangaroo habitat suitability}
\format{A raster layer with 35 rows, 36 columns, and 1260 cells:
\describe{
  \item{resolution}{size of cells in meters}
  \item{extent}{cartesian boundaries in meters}
  \item{coord. ref.}{spatial projection of layer}
}}
\usage{
egk_hab
}
\description{
A raster layer containing the predicted relative habitat suitability
for the Eastern Grey Kangaroo.
}
\keyword{datasets}
