% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_k}
\alias{egk_k}
\title{Eastern Grey Kangaroo carrying capacity}
\format{A raster layer with 35 rows, 36 columns, and 1260 cells:
\describe{
  \item{resolution}{size of cells in meters}
  \item{extent}{cartesian boundaries in meters}
  \item{coord. ref.}{spatial projection of layer}
}}
\usage{
egk_k
}
\description{
A raster stack containing the total number of Eastern Grey Kangaroos
each grid cell can support.
}
\keyword{datasets}
