% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_mat}
\alias{egk_mat}
\title{Eastern Grey Kangaroo life-stage transition matrix}
\format{A matrix with 3 rows and 3 columns:
\describe{
  \item{juvenile}{this column specifies the survival of juveniles into the next stage}
  \item{subadult}{this column specifies the survival of subadults into the next stage}
  \item{adult}{this column specifies the survival of adults into the next year and the number of offsring produced (first row)}
}}
\usage{
egk_mat
}
\description{
A matrix containing the survival and fecundity of Eastern Grey Kangaroos
at each of three life-stages - juvenile, subadult, and adult.
}
\keyword{datasets}
