% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_pop}
\alias{egk_pop}
\title{Eastern Grey Kangaroo populations}
\format{A raster stack with 35 rows, 36 columns, 1260 cells, and 3 layers:
\describe{
  \item{resolution}{size of cells in meters}
  \item{extent}{cartesian boundaries in meters}
  \item{coord. ref.}{spatial projection of layer}
}}
\usage{
egk_pop
}
\description{
A raster stack containing intital populations for each life-stage
of the Eastern Grey Kangaroo.
}
\keyword{datasets}
