% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egk_state}
\alias{egk_state}
\title{Eastern Grey Kangaroo state object}
\format{A state object with:
\describe{
  \item{population}{spatial layers describing the population}
  \item{habitat}{spatial layers describing habitat suitability and carrying capacity}
  \item{demography}{matrices describing survival and fecundity}
}}
\usage{
egk_state
}
\description{
An object containing the habitat, demographic and populations of Eastern Grey Kangaroos.
This is created with the \link[steps]{state} function.
}
\keyword{datasets}
