% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{envir_stochast}
\alias{envir_stochast}
\title{environmental stochastic function in C++}
\usage{
envir_stochast(tmat, matsd, equalsign = TRUE)
}
\arguments{
\item{tmat}{NumericMatrix. Vector with the initial abundance of each stage.}

\item{matsd}{NumericMatrix. Transition matrix.}

\item{equalsign}{bool. Should the environmental deviations have all the same sign and magnitude? TRUE or FALSE}
}
\description{
environmental stochastic function in C++
}

