% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-class.R
\name{order_events}
\alias{order_events}
\alias{schedule_event}
\alias{do_event}
\alias{.first}
\alias{.highest}
\alias{.last}
\alias{.lowest}
\alias{.normal}
\title{event class for dynamic metapopulation class models.}
\usage{
order_events()

schedule_event()

do_event()

.first()

.highest()

.last()

.lowest()

.normal()
}
\value{
A numeric.
}
\description{
This class is a way to order, schedule and run events which will have
an effect on the dynamic metapopulation models. Events can be temporally assigned. 
But need to be ordered with respect to metapopulation steps.
Events can have different temporal requirements. For example, management and demographic events
might be on a yearly scale, were fire might be weekly.

Event order
Assign event order: 1 = first (highest); 5 = normal; 10 = last (lowest).
}
