% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_results-class.R
\name{experiment}
\alias{experiment}
\title{Run an experiment to make spatially-explicit population projections}
\usage{
experiment(state, dynamics, timesteps = 100)
}
\arguments{
\item{state}{A state object - static habitat, population, and demography in a timestep}

\item{dynamics}{A dynamics object - modules that change habitat, population, and demography during and experiment}

\item{timesteps}{Number of timesteps used in the experiment}
}
\value{
An object of class \code{experiment_results}
}
\description{
Run an experiment to make spatially-explicit population projections
}
\examples{

library(raster)
library(dhmpr)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
test_demography <- build_demography(fake_transition_matrix(4), rlnorm(1))
test_population <- build_population(stack(replicate(4, test_habitat$carrying_capacity * 0.2)))
test_state <- build_state(test_habitat, test_demography, test_population)
fast_approximation <- build_dynamics(no_habitat_dynamics, no_demographic_dynamics, fast_population_dynamics)
results <- experiment(test_state, fast_approximation, timesteps = 10)
}
