% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_spread.R
\name{fire_spread}
\alias{fire_spread}
\title{fire spread object for dynamic habitat meta-populations.}
\usage{
fire_spread(habitat, fire_start_location = NA_real_, prob = 0.23,
  continue_to_burn_prob = 0, max_cells = 100000000L, directions = 8L,
  iterations = 1000000L, return_table = FALSE, id = FALSE,
  spread_prob_after_ignition = NA_real_, fire_history = NA)
}
\description{
This is an example function which we can import as a module to manipulate the habitat. 
This function uses cellular automata to spread fire across the habitat. These burnt cells can be used to reduce (or increase)
the probability of occurrence within the \code{habitat} object and thus re-shape or change patch suitability. 
Which will have an effect on the metapopulation models.
}

