% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{grad_huber_loss}
\alias{grad_huber_loss}
\title{Huber loss gradient}
\usage{
grad_huber_loss(a, h = 0.5)
}
\arguments{
\item{a}{vectore of gradients}

\item{h}{space parameter}
}
\value{
The huber loss gradient
}
\description{
Huber loss gradient
}
\examples{
grad_huber_loss(c(-0.5, 0.2, 1, 2))
}
