% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{huber_loss}
\alias{huber_loss}
\title{Huber loss}
\usage{
huber_loss(a, h = 0.5)
}
\arguments{
\item{a}{a number to calculate huber loss of}

\item{h}{a positive threshold for fine. Default is 0.5.}
}
\value{
The function returns the huber loss value of given a/
}
\description{
Huber loss
}
\details{
This function is slightly different from the regular huber loss, as all values smaller than -h
}
\examples{
x <- runif(1000, -5, 5)
(h_loss <- huber_loss(x, h = 0.5))
plot(h_loss ~ x)
}
