% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{is.demography}
\alias{is.demography}
\title{Verify demography object}
\usage{
is.demography(x)
}
\arguments{
\item{x}{a demography object to summarise}
}
\description{
Verify demography object
}
\examples{
is.demography(demo)
}
