% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{is.habitat}
\alias{is.habitat}
\title{Verify habitat object}
\usage{
is.habitat(x)
}
\arguments{
\item{x}{an object to print or test as an habitat object}
}
\description{
Verify habitat object
}
\examples{
is.habitat(test_habitat)
}
