% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{is.population}
\alias{is.population}
\title{Verify a population object}
\usage{
is.population(x)
}
\arguments{
\item{x}{an object to print or test as a population object}
}
\description{
Verify a population object
}
\examples{

is.population(pops)

}
