% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics-class.R
\name{is.population_dynamics}
\alias{is.population_dynamics}
\title{Check population dynamics object}
\usage{
is.population_dynamics(x)
}
\arguments{
\item{x}{a population_dynamics object}
}
\description{
Check population dynamics object
}
