% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results-class.R
\name{is.simulation_results}
\alias{is.simulation_results}
\title{Check simulation object}
\usage{
is.simulation_results(x)
}
\arguments{
\item{x}{a simulation_results objec}
}
\description{
Check simulation object
}
