% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{K}
\alias{K}
\title{#' @rdname habitat
#' @name K
#' @param x a raster of species habitit suitability (occupancy)
#' @param a parameter for carrying capacity model
#' @param b parameter for carrying capacity model
#' @param c parameter for carrying capacity model
#' @param type model form for converting occurrence to carrying capacity.
#' @author Skipton Woolley}
\usage{
K(x, a = 1, b = 0.2, c = 0.5, type = c("exp", "logit", "linear"))
}
\description{
#' @rdname habitat
#' @name K
#' @param x a raster of species habitit suitability (occupancy)
#' @param a parameter for carrying capacity model
#' @param b parameter for carrying capacity model
#' @param c parameter for carrying capacity model
#' @param type model form for converting occurrence to carrying capacity.
#' @author Skipton Woolley
}

