% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{linear_obj_term}
\alias{linear_obj_term}
\title{Linear Model Objective Function}
\usage{
linear_obj_term(pars, y.hat.fold, h = 0.5)
}
\arguments{
\item{pars}{Linear model equation's parameters.}

\item{y.hat.fold}{a table with two columns, and a row per pair we know its ordering. Each row includes the predictions for each pair.}

\item{h}{sapce parameter}
}
\value{
The function returns the value of the linear DSS model objective function.
}
\description{
Linear Model Objective Function
}
\details{
The objective function is of the shape:
\deqn{\frac{1}{2}\norm{w} + \frac{1}{2}*\sum_{pairs\in B}L_{h}\left( 1 - \hat{g}(x_{i}) - \hat{g}(x_{j}) \right)}
}
