% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{list2habitat}
\alias{list2habitat}
\title{#' @rdname habitat
#' @export
`area<-` <- function (habitat, new_area) {
  stopifnot(is.habitat(habitat))
  area_check(new_area)
  habitat[which(sapply(habitat,attr,"habitat")=='area')]<-new_area
  habitat
}}
\usage{
list2habitat(input)
}
\description{
#' @rdname habitat
#' @export
`area<-` <- function (habitat, new_area) {
  stopifnot(is.habitat(habitat))
  area_check(new_area)
  habitat[which(sapply(habitat,attr,"habitat")=='area')]<-new_area
  habitat
}
}
