% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metapop}
\alias{metapop}
\title{C++ metapopulation function for a single timestep.}
\usage{
metapop(presence, dist_mat, Ei, y)
}
\arguments{
\item{presence}{NumericVector Initial occupancy of each patch}

\item{dist_mat}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{Ei}{patch extinction rate at time i. note: In the future I need to pull this from demographic model.}

\item{y}{incidence function parameters.}
}
\description{
C++ metapopulation function for a single timestep.
}

