% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation-class.R
\name{metapopulation}
\alias{metapopulation}
\alias{plot.metapopulation}
\alias{pop_patch_name}
\title{metapopulation objects}
\usage{
metapopulation(nrep = 10, time = 20, dist, area, presence, y1 = 1,
  x1 = 1, e1 = 1, alpha = 1, beta = 1, disp_fun = "H",
  locations = NULL, c = 1, coln_fun = "H")

\method{plot}{metapopulation}(x, ...)

pop_patch_name(population)
}
\arguments{
\item{nrep}{Number of simulations.}

\item{time}{Number of time steps}

\item{dist}{Distances between patches (symetrical matrix)}

\item{area}{Area of patches - This needs to be calculated somehow - using occupancy models?}

\item{presence}{Initial occupancies of patches. Must be presence 1 or absence 0.}

\item{y1}{incidence function parameters}

\item{x1}{incidence function parameters}

\item{e1}{Minimum area of patches}

\item{alpha}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{beta}{double parameter that represents the shape of the dispersal kernel.}

\item{disp_fun}{char Which dispersal function to use. 'H' uses hanski(1994), if 'S' uses shaw(1995).}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}

\item{c}{double colonisation scale parameter see Ovaskainen 2002.}

\item{coln_fun}{char which colonisation function to use. 'H' = Hanski (1994), 'M'= Moilanen (2004), 'O'= Ovaskainen (2002). See decription for details.}

\item{x}{a metapopulation object}

\item{...}{other plot arguments}

\item{population}{matrix of states as cols and patches as rows.}
}
\description{
do some metapopulation modelling in R
}
\examples{
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
locations <- meta_data[,c('x1','x2')]
mp <- metapopulation(nrep=10, time=100, dist=dist, area=area, presence=presence,locations=locations,
                  x1 = 0.42, e1 = 0.0061, y1 = 1.2)
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- metapopulation(nrep=100, time=50, dist, area, presence, x = 0.42, e = 0.061, y = 1.2)
plot(mp)
}
\author{
Skipton Woolley

Skipton Woolley
}
\seealso{
\code{link{metapop_n_cpp}}
}

