% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-class.R
\name{module}
\alias{module}
\title{module objects}
\description{
Module for altering the habitat, they maybe fire spread, management, ect. 
Module sets up internal or custom functions to work with \code{habitat} and \code{simulation} objects.
}

