% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{non_linear_loss_fun}
\alias{non_linear_loss_fun}
\title{Non-linear Model Objective Function}
\usage{
non_linear_loss_fun(Fx, setB, setW, setR, lambdaB, lambdaW, lambdaR, h)
}
\arguments{
\item{Fx}{y}

\item{setB}{set B}

\item{setW}{set W}

\item{setR}{set R}

\item{lambdaB}{lambdaB}

\item{lambdaW}{lambdaW}

\item{lambdaR}{lambdaR}

\item{h}{space parameter}
}
\value{
The function returns the value of the linear DSS model objective function.
}
\description{
Non-linear Model Objective Function
}
\details{
The objective function is of the shape:
\deqn{\frac{1}{2}\norm{w} + \frac{1}{2}*\sum_{pairs\in O}L_{h}\left( 1 - \hat{g}(x_{i}) - \hat{g}(x_{j}) \right)}
}
\examples{
print("write an example")
}
