% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{obj_term_per_step}
\alias{obj_term_per_step}
\title{Objective function per tree iteration}
\usage{
obj_term_per_step(
  par,
  Fx_iter,
  tree,
  tree_data,
  setB,
  setW,
  setR,
  lambdaB,
  lambdaW,
  lambdaR,
  h = 0.5
)
}
\arguments{
\item{par}{par}

\item{Fx_iter}{Fx_iter}

\item{tree}{tree}

\item{tree_data}{tree_data}

\item{setB}{setB}

\item{setW}{setW}

\item{setR}{setR}

\item{lambdaB}{lambdaB}

\item{lambdaW}{lambdaW}

\item{lambdaR}{lambdaR}

\item{h}{h = 0.5}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Objective function per tree iteration
}
