% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.demographic_mod.R
\name{plot.demographic_mod}
\alias{plot.demographic_mod}
\title{Plot a demographic_mod object}
\usage{
\method{plot}{demographic_mod}(x, mean_pop = TRUE, ...)
}
\arguments{
\item{x}{a demographic_mod object}

\item{mean_pop}{logical. If TRUE plots the mean demographic change in population through time. Otherwise it plots the population changes for each stage in the transition matrix.}

\item{...}{other plot arguments}
}
\description{
Plot a demographic_mod demographic projection
}
\examples{
tmat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
matsd <- tmat/10
v0 <- c(80,20,0)
sim_t20_rep100 <- demographic_mod(v0=v0,tmat=tmat,matsd=matsd,estdem=TRUE,time=20,nrep=100) 
plot(sim_t20_rep100,mean_pop=TRUE)
plot(sim_t20_rep100,mean_pop=FALSE)
}
\author{
Skipton Woolley
}

