% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{plot.demography}
\alias{plot.demography}
\title{Plot details of demography object}
\usage{
\method{plot}{demography}(x, ...)
}
\arguments{
\item{x}{a demography object to summarise}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot details of demography object
}
\examples{
plot(demo)
}
