% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dlmp.R
\name{plot.dlmp}
\alias{plot.dlmp}
\title{Plot a dlmp object}
\usage{
\method{plot}{dlmp}(x, mean_pop = TRUE, ...)
}
\arguments{
\item{x}{a dlmp object}

\item{mean_pop}{logical. If TRUE plots the mean demographic change in population through time. Otherwise it plots the population changes for each stage in the transition matrix.}

\item{...}{other plot arguments}
}
\description{
Plot a dlmp demographic projection
}
\examples{
tmat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
matsd <- tmat/10
v0 <- c(80,20,0)
sim_t20_rep100 <- dlmp(v0=v0,tmat=tmat,matsd=matsd,estdem=TRUE,time=20,nrep=100) 
plot(sim_t20_rep100,mean_pop=TRUE)
plot(sim_t20_rep100,mean_pop=FALSE)
}
\author{
Skipton Woolley
}

