% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_results-class.R
\name{plot.experiment_results}
\alias{plot.experiment_results}
\title{Plot an experiment object}
\usage{
\method{plot}{experiment_results}(results, object = "population",
  type = "raster", stage = NULL, ...)
}
\arguments{
\item{results}{}

\item{object}{}

\item{type}{}

\item{...}{}
}
\description{
Plot an experiment object
}
