% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.metapop_mod.R
\name{plot.metapop_mod}
\alias{plot.metapop_mod}
\title{Plot a metapop_mod object}
\usage{
\method{plot}{metapop_mod}(x, ...)
}
\arguments{
\item{x}{a metapop_mod object}

\item{...}{other plot arguments}
}
\description{
Plot a metapop_mod object.
}
\examples{
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- metapop_mod(nrep=100, time=50, dist, area, presence, x = 0.42, e = 0.061, y = 1.2)
plot(mp)
}
\author{
Skipton Woolley
}

