% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim_metapop.R
\name{plot.sim_metapop}
\alias{plot.sim_metapop}
\title{Plot a sim_metapop object}
\usage{
\method{plot}{sim_metapop}(x, ...)
}
\arguments{
\item{x}{a sim_metapop object}

\item{...}{other plot arguments}
}
\description{
Plot a sim_metapop object.
}
\examples{
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- sim_metapop(nrep=100, time=50, dist, area, presence, x = 0.42, e = 0.061, y = 1.2)
plot(mp)
}
\author{
Skipton Woolley
}

