% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonizer.R
\name{polygonizer}
\alias{polygonizer}
\title{polygonizer
Convert raster data to a ESRI polygon shapefile and (optionally) a SpatialPolygonsDataFrame}
\usage{
polygonizer(x, outshape = NULL, pypath = NULL, readpoly = TRUE,
  quietish = TRUE)
}
\arguments{
\item{x}{an R Raster layer, or the file path to a raster file recognised by GDAL}

\item{outshape}{the path to the output shapefile (if NULL, a temporary file will 
be created)}

\item{pypath}{the path to gdal_polygonize.py or OSGeo4W.bat (if NULL, the function 
will attempt to determine the location)}

\item{readpoly}{should the polygon shapefile be read back into R, and returned by
this function? (logical)}

\item{quietish}{should (some) messages be suppressed? (logical)}
}
\description{
polygonizer
Convert raster data to a ESRI polygon shapefile and (optionally) a SpatialPolygonsDataFrame
}
\author{
John Baumgartner
}

