% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{population}
\alias{population}
\alias{is.population}
\alias{print.population}
\title{Create a population object to use in a state object.}
\usage{
population(population_raster)

is.population(x)

\method{print}{population}(x, ...)
}
\arguments{
\item{population_raster}{A raster stack (grid cell-based) with one layer
for each life stage.}

\item{x}{A population object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{population}
}
\description{
A population object contains information on how populations
(abundances of individuals) change in space and time.
}
\details{
A population object is a sub-component of a \link[steps]{state} object
and is modified in each timestep of a simulation. During a simulation,
a population object tracks changes in total individuals per grid cells
based on population dynamic functions selected or created by the user.
}
\examples{

library(steps)
library(raster)

# Construct the population object
pop <- population(egk_pop)

# Test if object is of the type 'population'
is.population(test_population)

# Print information about the 'population' object
print(test_population)
}
