% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{prediction.STEPS}
\alias{prediction.STEPS}
\title{DSS prediction}
\usage{
prediction.STEPS(
  tree_model,
  newdata,
  shrinkage = 0.1,
  rho = NULL,
  summarise_fun = mean
)
}
\arguments{
\item{tree_model}{list of trees output. Must include arguments with the names "weights" an
"trees_info". even if it is just one model it should still be a list}

\item{newdata}{newdata set}

\item{shrinkage}{shrinakge rate, default is 0.1}

\item{rho}{if NULL than the weights per iteration are taken from tree_model$weights}

\item{summarise_fun}{if the models are from cross validation, or we have more than one,
then we can summarise the different predictions per observations by mean or emdian for example.
Default is set to mean.}
}
\value{
A vector with predictions
}
\description{
DSS prediction
}
\examples{
print("missing ecample")
}
