% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{print.demography}
\alias{print.demography}
\title{Print details of a demography object}
\usage{
\method{print}{demography}(x, ...)
}
\arguments{
\item{x}{an object to print or test as an demography object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a demography object
}
