% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-class.R
\name{print.demography_dynamics}
\alias{print.demography_dynamics}
\title{Print details of a demography_dynamics object}
\usage{
\method{print}{demography_dynamics}(x, ...)
}
\arguments{
\item{x}{an object to print or test as an demography_dynamic object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a demography_dynamics object
}
