% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics-class.R
\name{print.dynamics}
\alias{print.dynamics}
\title{Print details of a dynamics object}
\usage{
\method{print}{dynamics}(x, ...)
}
\arguments{
\item{x}{an object to print or test as a dynamics object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a dynamics object
}
