% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_results-class.R
\name{print.experiment_results}
\alias{print.experiment_results}
\title{Print details of a experiment object}
\usage{
\method{print}{experiment_results}(results)
}
\arguments{
\item{results}{an object to print or test as an experiment_results object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a experiment object
}
