% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{print.habitat}
\alias{print.habitat}
\title{Print details of a habitat object}
\usage{
\method{print}{habitat}(x, ...)
}
\arguments{
\item{x}{an object to print or test as an habitat object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a habitat object
}
