% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-class.R
\name{print.habitat_dynamics}
\alias{print.habitat_dynamics}
\title{Print details of a habitat_dynamics object}
\usage{
\method{print}{habitat_dynamics}(x, ...)
}
\arguments{
\item{x}{an object to print or test as an habitat_dynamic object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a habitat_dynamics object
}
