% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{print.population}
\alias{print.population}
\title{Print details of a population object}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{an object to print or test as a population object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a population object
}
