% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics-class.R
\name{print.population_dynamics}
\alias{print.population_dynamics}
\title{Print details of a population_dynamics object}
\usage{
\method{print}{population_dynamics}(x, ...)
}
\arguments{
\item{x}{an object to print or test as an population_dynamic object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a population_dynamics object
}
