% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-class.R
\name{print.state}
\alias{print.state}
\title{Print details of a state object}
\usage{
\method{print}{state}(x, ...)
}
\arguments{
\item{x}{an object to print or test as a state object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print details of a state object
}
