% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal-class.R
\name{probdisp}
\alias{probdisp}
\title{#' @rdname dispersal
#' @param x an object to print or test as a transition object
#' @param \dots further arguments passed to or from other methods.
#' @export
#' @examples
#' # print method
#' print(dp)
#'
print.dispersal <- function(x,...){
  disp_info <- list()
  for (i in base::seq_along(x)[-length(x)]) {
  disp_info[[i]]  <- base::paste(
      base::names(x[[i]]),
      base::format(x[[i]], scientific = FALSE),
      sep = " = ",
      collapse = "\n "
    ) 
  };
  if(length(disp_info)==2) text <- sprintf('dispersal function with disperal kernal of:\n %s\n and probability of dispersal of:\n %s\n for stages.',disp_info[[1]],disp_info[[2]])
  else text <- sprintf('dispersal function with disperal kernal of:\n %s\n for stages.',disp_info[[1]])
  cat(text)
}}
\usage{
probdisp(disp, habitat)
}
\description{
#' @rdname dispersal
#' @param x an object to print or test as a transition object
#' @param \dots further arguments passed to or from other methods.
#' @export
#' @examples
#' # print method
#' print(dp)
#'
print.dispersal <- function(x,...){
  disp_info <- list()
  for (i in base::seq_along(x)[-length(x)]) {
  disp_info[[i]]  <- base::paste(
      base::names(x[[i]]),
      base::format(x[[i]], scientific = FALSE),
      sep = " = ",
      collapse = "\n "
    ) 
  };
  if(length(disp_info)==2) text <- sprintf('dispersal function with disperal kernal of:\n %s\n and probability of dispersal of:\n %s\n for stages.',disp_info[[1]],disp_info[[2]])
  else text <- sprintf('dispersal function with disperal kernal of:\n %s\n for stages.',disp_info[[1]])
  cat(text)
}
}

