\name{ramasr-package}
\alias{ramasr-package}
\alias{ramasr}
\docType{package}
\title{
\packageTitle{ramasr}
}
\description{
\packageDescription{ramasr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ramasr}
\packageIndices{ramasr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ramasr}

Maintainer: \packageMaintainer{ramasr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
