% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{rank_init_fun}
\alias{rank_init_fun}
\title{Initial ranking function}
\usage{
rank_init_fun(x, ord.pairs)
}
\arguments{
\item{x}{a data.frame or a matrix that holds one row per observation.
each row represents an index of an observation to look for in the \code{pairs.tab}}

\item{ord.pairs}{a table with sorted pairs of observations.}
}
\value{
The function counts for each observation how many times it was smaller than
         other observations and how many times it was greater.
}
\description{
Initial ranking function
}
\details{
As a good starting point for y values we will use the information we have on our sorted pairs.
         An observations will get a higher rank for starter if they appear a lot of times in a
        severe state comapring to others observations.
}
\examples{
set.seed(6324)
}
