% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmp.metapop.R
\name{sim_metapop}
\alias{sim_metapop}
\title{sim_metapop}
\usage{
sim_metapop(nrep = 10, time = 20, dist, area, presence, y = 1, x = 1,
  e = 1, alpha = 1, beta = 1, hanski_dispersal_kernal = TRUE,
  locations = NULL)
}
\arguments{
\item{nrep}{Number of simulations.}

\item{time}{Number of time steps}

\item{dist}{Distances between patches (symetrical matrix)}

\item{area}{Area of patches - This needs to be calculated somehow - using occupancy models?}

\item{presence}{Initial occupancies of patches. Must be presence 1 or absence 0.}

\item{y}{incidence function parameters}

\item{x}{incidence function parameters}

\item{e}{Minimum area of patches}

\item{alpha}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{beta}{double parameter that represents the shape of the dispersal kernel.}

\item{hanski_dispersal_kernal}{bool if true uses hanski(1994), if false uses shaw(1995).}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}
}
\description{
Demographic projection
}
\details{
Make deterministic and stochastic demographic projections according to a transition matrix.

Simulate a metapopulation system in C++
this function is not complete.
}
\examples{
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- sim_metapop(nrep=10, time=20, dist, area, presence, x = 0.42, e = 0.061, y = 15.2)
}
\author{
Skipton Woolley
}
\seealso{
\code{link{metapop_n_cpp}}
}

