% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results.R
\name{simulation}
\alias{simulation}
\alias{is.simulation_results}
\title{Run an simulation experiment to make many spatially-explicit population projections and look at the variation.}
\usage{
simulation(state, dynamics, timesteps, simulations, check = TRUE, ...)

is.simulation_results(results)
}
\arguments{
\item{state}{a state object - static habitat, population, and demography in a timestep}

\item{dynamics}{a dynamics object - modules that change habitat, population, and demography during and experiment}

\item{timesteps}{number of timesteps used in the experiment}

\item{simulations}{an experiment_reults object}
}
\value{
An object of class \code{simulation_results}
}
\description{
A habitat object is used to store spatially-explicit information on habitat suitability and the carrying_capacity of a landscape.
It is a sub-component of a \link[steps]{state} object and is modified in each timestep of an experiment.
}
\examples{

library(steps)
library(raster)

r <- raster(system.file("external/test.grd", package="raster"))

test_habitat <- build_habitat(habitat_suitability = r / cellStats(r, "max"), carrying_capacity = ceiling(r * 0.1))
test_demography <- build_demography(transition_matrix = steps:::fake_transition_matrix(4), dispersal_parameters = rlnorm(1))
test_population <- build_population(stack(replicate(4, test_habitat$carrying_capacity * 0.2)))
test_state <- build_state(test_habitat, test_demography, test_population)
fast_approximation <- build_dynamics(steps:::no_habitat_dynamics, steps:::no_demographic_dynamics, steps:::fast_population_dynamics)
simulation_results <- simulation(test_state,fast_approximation,timesteps=10,simulations=10)

# Test if object is of the type 'experiment results'
  
is.simulation_results(results)
}
