% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-class.R
\name{state}
\alias{state}
\alias{is.state}
\alias{print.state}
\title{Create a state object}
\usage{
state(population, habitat, demography)

is.state(x)

\method{print}{state}(x, ...)
}
\arguments{
\item{population}{A \link[steps]{population} object.}

\item{habitat}{A \link[steps]{habitat} object.}

\item{demography}{A \link[steps]{demography} object.}

\item{x}{A state object to print or test.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{state}
}
\description{
A state object is a static representation of habitat, population,
and demography in a single timestep.
}
\details{
A state object is modified in each timestep of a simulation based on
specified functions in a \link[steps]{dynamics} object. When first
created, the state object will hold initial values for habitat, population,
and demography which may or may not change throughout a simulation.
}
\examples{

library(steps)
library(raster)

# Create habitat, demography, and population objects
hab <- habitat(egk_hab, egk_k)
dem <- demography(egk_mat)
pop <- population(egk_pop)

# Construct the state object
state <- state(pop, hab, dem)

# Test if object is of the type 'state'
is.state(test_state)

# Print information about the 'state' object
print(test_state)
}
