% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{stop_term_fun}
\alias{stop_term_fun}
\title{Stopping terms validation}
\usage{
stop_term_fun(iter_count, stop_terms)
}
\arguments{
\item{iter_count}{iter_count}

\item{stop_terms}{stop_terms}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Stopping terms validation
}
