% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{summary.demography}
\alias{summary.demography}
\title{Show details of demography object}
\usage{
\method{summary}{demography}(x, ...)
}
\arguments{
\item{x}{a demography object to summarise}

\item{...}{further arguments passed to or from other methods}
}
\description{
Show details of demography object
}
\examples{
summary(demo)
}
