% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix_functions.R
\name{trans_matrix}
\alias{as.trans_matrix}
\alias{is.trans_matrix}
\alias{trans_matrix}
\title{transition matrix objects}
\usage{
as.trans_matrix(x, names.st = NULL, ...)

is.trans_matrix(x)
}
\arguments{
\item{x}{For as.trans_matrix, x is a square matrix, that has transition states between population stages.}

\item{x}{a trans_matrix object}
}
\value{
An object of class tmatrix, i.e, resulting from as.trans_matrix.
}
\description{
Transition population matrix functions
}
\examples{
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.trans_matrix(mat)
is.trans_matrix(tmat)
}
\author{
Skipton Woolley
}

