% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-class.R
\name{as.transition}
\alias{as.transition}
\alias{is.transition}
\alias{plot.transition}
\alias{summary.transition}
\title{transition objects}
\usage{
as.transition(x, ...)

\method{summary}{transition}(x, ...)

\method{plot}{transition}(x, ...)

is.transition(x)
}
\arguments{
\item{x}{For as.transition, x is a square matrix, that has transition states between population stages.}

\item{...}{other function calls.}

\item{object}{an object of \code{transition} class}
}
\value{
An object of class tmatrix, i.e, resulting from as.transition.
}
\description{
prints the main parameters of the transition matrix: the finite rate of increase ("lambda"), the stable stage distribution, the reproductive value and the sensitivities and elasticities matrices.
}
\examples{
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
rownames(mat) <- colnames(mat) <- c('larval','juvenile','adult')
tmat <- as.transition(mat)
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition(mat)
summary(tmat)
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition(mat)
plot(tmat)
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition(mat)
is.transition(tmat)
}
\author{
Skipton Woolley

Nick Golding
}

