% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_transition_functions-class.R
\name{transition_function}
\alias{transition_function}
\title{Create a growth transition function}
\description{
A growth transition function defines how spatial objects or custom functions influence
survival and fecundity. A user may select from built-in functions or provide a custom written function
to modify survival and fecundity throughout a simulation.
}
\details{
Two built-in functions are provided for the user to select, however, a user may also provide
other custom written density dependence functions.
}
\seealso{
\itemize{
  \item{\link[steps]{modified_transition} to use rasters to modify survival and fecundity}
  \item{\link[steps]{competition_density} to use relationship to carrying capacity to modify
  survival and fecundity}
  }
}
