% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix-class.R
\name{transition_matrix}
\alias{as.transition_matrix}
\alias{is.transition_matrix}
\alias{plot.transition_matrix}
\alias{summary.transition_matrix}
\alias{transition_matrix}
\title{transition matrix objects}
\usage{
as.transition_matrix(object, names.st = NULL, ...)

is.transition_matrix(object)

\method{summary}{transition_matrix}(object, ...)

\method{plot}{transition_matrix}(object, ...)
}
\arguments{
\item{object}{For as.transition_matrix, x is a square matrix, that has transition states between population stages.}

\item{...}{other function calls from igraph.}

\item{object}{a transition_matrix object}

\item{object}{a transition_matrix object}

\item{object}{a transition_matrix object}
}
\value{
An object of class tmatrix, i.e, resulting from as.transition_matrix.
}
\description{
prints the main parameters of the transition matrix: the finite rate of increase ("lambda"), the stable stage distribution, the reproductive value and the sensitivities and elasticities matrices.
}
\examples{
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition_matrix(mat)
is.transition_matrix(tmat)
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition_matrix(mat)
summary(tmat)
mat <- matrix(c(.53,0,.42,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
tmat <- as.transition_matrix(mat)
plot(tmat)
}
\author{
Skipton Woolley

Nick Golding
}

