% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{var_error_term}
\alias{var_error_term}
\title{Variance error term}
\usage{
var_error_term(subject.Fx)
}
\arguments{
\item{subject.Fx}{subject.Fx}
}
\value{
The function returns variance error term
}
\description{
Variance error term
}
\details{
The
}
\examples{
var_error_term(c(1, 2, 3))
}
