% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dispersal-functions.R
\name{cellular_automata_dispersal}
\alias{cellular_automata_dispersal}
\title{Cellular automata dispersal}
\usage{
cellular_automata_dispersal(max_cells = Inf,
  dispersal_proportion = set_proportion_dispersing(),
  barriers_map = NULL, use_suitability = TRUE,
  carrying_capacity = "carrying_capacity")
}
\arguments{
\item{max_cells}{the maximum number of cell movements that each individual in
each life stage can disperse in whole integers.}

\item{dispersal_proportion}{proportions of individuals (0 to 1) that can
disperse in each life stage. Can also be a custom function that relates
the proportion dispersing to a spatial layer in the landscape object
(e.g. carrying capacity).}

\item{barriers_map}{the name of a spatial layer in the landscape object that
contains cell values between 0 (no barrier) and 1 (full barrier) Any
values between 0 and 1 indicate the permeability of the barrier.}

\item{use_suitability}{should habitat suitability be used to control the
likelihood of individuals dispersing into cells? The default is TRUE. Note,
if a barrier map is also provided, the suitability map is multiplied with
the barrier map to generate a permeability map of the landscape.}

\item{carrying_capacity}{the name of a spatial layer in the landscape object
that specifies the carrying capacity in each cell.}
}
\description{
The cellular_automata_dispersal function modifies populations
using rule-based cell movements. This function allows the use
of barriers in the landscape to influence dispersal. The
function is computationally efficient, however, because
individuals are dispersed, performance scales with the
population sizes in each cell across a landscape.
}
\examples{

test_ca_dispersal <- cellular_automata_dispersal()
}
