% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_proj_n.R
\name{demo_proj_n}
\alias{demo_proj_n}
\title{Demographic projection}
\usage{
demo_proj_n(v0, tmat, matsd = NULL, stmat = NULL, estamb = FALSE,
  estdem = FALSE, equalsign = TRUE, fecundity1 = FALSE, nrep = 10,
  time = 10)
}
\arguments{
\item{v0}{}
}
\value{
list of demograph projections
}
\description{
Demographic projection
}

