% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-class.R
\name{as.demography}
\alias{as.demography}
\alias{summary.demography}
\alias{stages}
\alias{plot.demography}
\alias{is.demography}
\alias{estimate_demography}
\title{demography objects}
\usage{
as.demography(x, method = "global", ...)

\method{summary}{demography}(x, ...)

stages(demography)

\method{plot}{demography}(x, ...)

is.demography(x)

estimate_demography(demography_object, habitat, parameters)
}
\arguments{
\item{x}{For as.demography, x is a square matrix, that has stage stages between population stages.}

\item{method}{can be 'global' or 'local' stage matrices. 'global' is the default and a single stage-based stage matrix 
will be used for all populations. if 'local' method is called a cell specific stage matrix will be setup for all non-NA cells
 in the underlying habitat_suitanility.}

\item{demography}{a demography object}

\item{object}{an object of \code{demography} class}

\item{demographic_params}{a list of parameters which can be used to manipulate demographic projections.
\itemize{
#'\item{stage_matrix_sd}{Matrix with the standard deviation of the probabilities in \code{mat}.}
}}
}
\value{
An object of class demography, i.e, resulting from as.demography.
}
\description{
demography is one of the main functions for dhmpr, it helps you construct and process stage based matricies.
Once a demography object is created \code{summary} will return:
 The finite rate of increase "lambda",
 the stable stage distribution,
 the reproductive value,
 and the sensitivities and elasticities matrices.
 \code{plot} will return a graph object that plots the stages between and amoungest each stage of the population matrix.
}
\examples{
mat <- matrix(c(.53,0,.52,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
colnames(mat) <- rownames(mat) <- c('larvae','juvenile','adult') 
demo <- as.demography(mat)
summary(tmat)

# get component stages
stages(demo)
plot(tmat)

is.demography(tmat)
}
\author{
Skipton Woolley

Nick Golding
}
