% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-class.R
\name{as.demography_dynamics}
\alias{as.demography_dynamics}
\alias{is.demography_dynamics}
\alias{print.demography_dynamics}
\alias{no_demographic_dynamics}
\alias{envstoch_demographic_dynamics}
\title{Change the demography in a state object}
\usage{
as.demography_dynamics(demography_dynamics_function)

is.demography_dynamics(x)

\method{print}{demography_dynamics}(x, ...)

no_demographic_dynamics(state, timestep)

envstoch_demographic_dynamics(global_transition_matrix, stochasticity = 0)
}
\arguments{
\item{demography_dynamics_function}{A function that operates on a state object to change demography at specified timesteps. User may enter a custom function or select a pre-defined module - see examples.}

\item{x}{an object to print or test as an demography_dynamic object}

\item{...}{further arguments passed to or from other methods}

\item{state}{a state object to apply the demographic function to}

\item{timestep}{the timestep in the experiment to apply the demographic function to the state object}

\item{global_transition_matrix}{a life-stage transition matrix}

\item{stochasticity}{a matrix with standard deviations (consistent or varying) around the transition means with matching dimensions as the life-stage transition matrix or a number representing a consitent standard deviation to apply to all transitions (default is 0)}
}
\value{
An object of class \code{demography_dynamics}
}
\description{
A 'demography dynamics' object is used to modify life-stage transition matrices - adding stochasticity for example.
It is a sub-component of a \link[steps]{dynamics} object and is executed in each timestep of an experiment.
}
\examples{

library(steps)
library(raster)

# create a generic function that simple returns an unmodified state object at each timestep

example_function <- function (state, timestep) {
state
}

define the function as a demography_dynamics object

no_demography_dynamics <- as.demography_dynamics(example_function)

# Test if object is of the type 'demography dynamics'
  
is.demography_dynamics(no_demography_dynamics)

print(no_demography_dynamics)
}
