% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_dynamics-class.R
\name{demography_dynamics}
\alias{demography_dynamics}
\alias{as.demography_dynamics}
\alias{is.demography_dynamics}
\alias{run_demography_dynamics}
\title{demography_dynamics objects}
\usage{
as.demography_dynamics(fun, params, check = FALSE, ...)

is.demography_dynamics(x)

run_demography_dynamics(demography_dynamics, demography_object, time_step, ...)
}
\description{
demography_dynamics are functions for altering the underlying demographic process in the \link[code]{experiment}. 
Demography_dynamics are functions which either directly affect the stage-based demographic matrix or other demographic processes excluding dispersal, any functions which alter dispersal are called from \link[dhmpr]{dispersal} or \link[dhmpr]{dispersal_dynamics}.
\code{demography_dynamics} sets up internal or custom functions to work with \code{demography} and \code{experiment} objects.

this bad boy will run the demography_dynamics in a experiment.
}
\examples{
## Create population
library(raster)
library(dhmpr)
#set a demography matrix
mat <- matrix(c(.53,0,.62,0.15,0.87,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
colnames(mat) <- rownames(mat) <- c('larvae','juvenile','adult') 
demog <- as.demography(mat)
n_stages <- length(states(demog))
}
