% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_proportion_functions-class.R
\name{density_dependence_dispersing}
\alias{density_dependence_dispersing}
\title{Density-dependent proportions of populations dispersing}
\usage{
density_dependence_dispersing()
}
\description{
The proportion of populations dispersing will be density dependent in a simulation. Proportions
of populations in each life stage dispering is adjusted based on available carrying capacity.
If life-stages are set by the \link[steps]{population_density_dependence_functions}, these
will be used to determine how close the population is to carrying capacity. If no
life-stages are set or density dependence is set to NULL in \link[steps]{population_dynamics},
the function will consider all life-stages in the calculation.
}
\examples{

# Example of a proportion function that disperses all populations based on their approach
# to carrying capacity

\dontrun{
prop_dispersal <- density_dependence_dispersing()

kb_dispersal <- kernel_dispersal(dispersal_proportion = prop_dispersal,
                      dispersal_kernel = exponential_dispersal_kernel(distance_decay = 1000))

ls <- landscape(population = egk_pop, suitability = egk_hab, carrying_capacity = egk_k)

pd <- population_dynamics(change = growth(egk_mat), dispersal = kb_dispersal)

simulation(landscape = ls, population_dynamics = pd, habitat_dynamics = NULL, timesteps = 20)
}
}
