% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_proportion_functions-class.R
\name{density_dependence_dispersing}
\alias{density_dependence_dispersing}
\title{Density-dependent proportions of populations dispersing}
\usage{
density_dependence_dispersing()
}
\description{
The proportion of populations dispersing to be density dependent. Proportions of
individuals in each life stage dispering is adjusted based on available carrying capacity.
If \code{density_dependence_stages} is set in \link[steps]{simulation}, these will be used
to determine how close the population is to carrying capacity. If no \code{density_dependence_stages}
are set, the function will consider all life-stages in the calculation.
}
\examples{

test_dispersal_function <- density_dependence_dispersing()
}
