% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_funs.R
\name{extent}
\alias{bcb}
\alias{dispersalFFT}
\alias{extend}
\alias{extent}
\alias{setupFFT}
\title{dispersal functions}
\usage{
extend(x, factor = 2)

bcb(x, y, f = I)

setupFFT(x, y, f, factor = 2)

dispersalFFT(popmat, fs)
}
\examples{
# small functions that should be in base R
seq_range <- function (range, by = 1) seq(range[1], range[2], by = by)
ifft <- function (z) fft(z, inverse = TRUE)

# coordinates of grid cells
# *don't make n too big, as the dense version will take forever!*
n <- 50 * c(2, 2)
y <- seq_len(n[1])
x <- seq_len(n[2])

# dispersal function acting on distance matrix
# cut-off dispersal at the minimum dimension of the grid 
f <- function (d, cutoff = min(n)) {
 ifelse (d > cutoff, 0, exp(-d))
}

# f <- function (d) exp(-d)
# initial population on grid (one stage)
pop <- matrix(rpois(length(x) * length(y),10),
             length(y), length(x))
             
# setup for the fft approach (run this once, before the simulation)
fs <- setupFFT(x = x, y = y, f = f, factor = 1)

# apply dispersal to the population (need to run this separately for each stage)
pop_new <- dispersalFFT(popmat = pop, fs = fs)
}

