% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal-class.R
\name{as.dispersal}
\alias{as.dispersal}
\alias{d}
\alias{is.dispersal}
\alias{print.dispersal}
\title{dispersal class for metapoplations}
\usage{
as.dispersal(params, ...)

d(params, ...)

is.dispersal(x)

\method{print}{dispersal}(x, ...)
}
\arguments{
\item{x}{an object to be tested as a dispersal transfun object}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
creates a function that governs dispersal capacity of stages in a population. The input is a list which contains 
 the first list is a dispersal kernal \code{alpha}  value for each stage, the second named list \code{probability} 
is the proportion of that stage will disperse. For example a probability of 0.2 for stage larvae
means a random 20% of larve will a try and disperse to patches, the distance they can disperse is governed 
by the dispersal kernal (alpha). If params = NULL, a dispersal kernal of 1 is given to all stages, 
and all stages will attempt to undertake dispersal.
}
\examples{
params <- list(alpha=list('larvae'=2,'juvenile'=0,'adult'=3),
               probability=list('larvae'=0.2,'juvenile'=0,'adult'=0.6))  
               
dp <- as.dispersal(params)

dp <-as.dispersal(NULL)
d(NULL)

is.dispersal(disp)
# print method
print(dp)

}

