% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal-class.R
\name{dispersal_core_fft}
\alias{dispersal_core_fft}
\title{# apply dispersal to the population (need to run this separately for each stage)}
\usage{
dispersal_core_fft(params, habitat)
}
\description{
# apply dispersal to the population (need to run this separately for each stage)
}

