% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_function-class.R
\name{exponential_dispersal_kernel}
\alias{exponential_dispersal_kernel}
\alias{print.dispersal_function}
\title{Create a dispersal function}
\usage{
exponential_dispersal_kernel(distance_decay = 0.5, normalize = FALSE)

\method{print}{dispersal_function}(x, ...)
}
\arguments{
\item{distance_decay}{a parameter to control the rate at which the population disperses with distance}

\item{normalize}{should the normalising constant be used - default is false.}

\item{x}{an object to print or test as a dispersal_function object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{dispersal_function}
}
\description{
A dispersal kernal function is mathematical representation of how species redistribute across the landscape. 
More common kernels are built-in for the user to select, however, a user may also provide a custom written dispersal kernel.
}
\examples{

test_dispersal_function <- exponential_dispersal_kernel()

print(test_dispersal_function)
}
