% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel_functions-class.R
\name{exponential_dispersal_kernel}
\alias{exponential_dispersal_kernel}
\title{Create a dispersal function}
\usage{
exponential_dispersal_kernel(distance_decay = 0.5, normalize = FALSE)
}
\arguments{
\item{distance_decay}{a parameter to control the rate at which the population disperses with distance}

\item{normalize}{should the normalising constant be used - default is false.}
}
\value{
An object of class \code{dispersal_function}
}
\description{
A dispersal kernal function is mathematical representation of how species redistribute across the landscape. 
More common kernels are built-in for the user to select, however, a user may also provide a custom written dispersal kernel.
}
\examples{

test_dispersal_function <- exponential_dispersal_kernel()
}
