% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_proportion_functions-class.R
\name{dispersal_proportion_function}
\alias{dispersal_proportion_function}
\title{Create a proportion dispersing function}
\description{
A proportion dispersing function generates the proportions of species
that disperse from cells based on landscape features.
}
\details{
The default \code{set_proportion_dispersing} function and parameters returns
full dispersal for all life stages. Additional proportion dispersing functions
are provided in the software for the user to select, however, a user may also
provide a custom written proportion dispersing function. Please see the tutorial
vignette titled "Creating custom *steps* functions" for information on how to
write custom functions for use in simulations.
}
\seealso{
\itemize{
  \item{\link[steps]{set_proportion_dispersing} controls the proportions of each life-stage that disperse}
  \item{\link[steps]{density_dependence_dispersing} proportions of dispersing populations are controlled by
  approach to carrying capacity}
  }
}
