% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-functions.R
\name{disturbance}
\alias{disturbance}
\title{Disturbance}
\usage{
disturbance(disturbance_layers, effect_time = 1)
}
\arguments{
\item{disturbance_layers}{the name of spatial layer(s) in the landscape object with disturbances used
to alter the habitat object for each timestep (number of layers must match the intended timesteps)}

\item{effect_time}{the number of timesteps that the disturbance layer will act on the habitat object}
}
\description{
Disturbance Description
}
\examples{

# Use the disturbance function to modify the habitat using spatial
# layers (stored in the landscape object and called "logging"):

logging <- disturbance(disturbance_layers = "logging",
                    effect_time = 1)
}
