% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamics-class.R
\name{dynamics}
\alias{dynamics}
\alias{is.dynamics}
\alias{print.dynamics}
\title{Create a dynamics object}
\usage{
dynamics(population_dynamics, habitat_dynamics, demography_dynamics,
  order = c("population_dynamics", "habitat_dynamics",
  "demography_dynamics"))

is.dynamics(x)

\method{print}{dynamics}(x, ...)
}
\arguments{
\item{population_dynamics}{A module to alter the population object in a simulation}

\item{habitat_dynamics}{A module to alter the habitat object in a simulation}

\item{demography_dynamics}{A module to alter the habitat object in a simulation}

\item{order}{The order to apply the dynamics at each timestep in a simulation}

\item{x}{an object to print or test as a dynamics object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{dynamics}
}
\description{
A dynamics object is a group of modular functions used modify habitat,
population, and demography in a timestep.
}
\details{
A dynamics object modifies a \link[steps]{state} object in each timestep
of a simulation based on specified habitat, demography, or population
dynamics functions.
}
\examples{

library(steps)
library(raster)

dynamics <- dynamics(population_dynamics(),
                     habitat_dynamics(),
                     demography_dynamics())

# Test if object is of the type 'dynamics'
  
is.dynamics(test_dynamics)

print(test_dynamics)
}
