% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dispersal-functions.R
\name{fast_dispersal}
\alias{fast_dispersal}
\title{Fast diffusion-based dispersal}
\usage{

  fast_dispersal(dispersal_kernel = exponential_dispersal_kernel(distance_decay
  = 0.1), dispersal_proportion = set_proportion_dispersing())
}
\arguments{
\item{dispersal_kernel}{a single built-in or user-defined distance dispersal
kernel function.}

\item{dispersal_proportion}{proportions of individuals (0 to 1) that can
disperse in each life stage. Can also be a custom function that relates
the proportion dispersing to a spatial layer in the landscape object
(e.g. carrying capacity).}
}
\description{
The fast_dispersal function uses kernel-based dispersal
to modify the population with a user-defined diffusion distribution
and a fast-fourier transformation (FFT) computational algorithm. It
is computationally efficient and very fast, however, only useful for
situations where dispersal barriers or arrival based on habitat or
carrying capacity are not required (e.g. a homogeneous landscape or
where diffusion alone is sufficient to explain dispersal patterns).
Dispersal is not constrained to suitable habitat or available carrying
capacity.
}
\examples{

test_fast_dispersal <- fast_dispersal()
}
