% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-functions.R
\name{fire_effects}
\alias{fire_effects}
\title{Fire}
\usage{
fire_effects(fire_layers, lag = 3,
  regeneration_function = function(time) {     -time })
}
\arguments{
\item{fire_layers}{the name(s) of spatial layer(s) in the landscape object with fire disturbances used
to alter the habitat object for each timestep (number of layers must match the intended timesteps)}

\item{lag}{the number of timesteps that the fire layer will act on the habitat object}

\item{regeneration_function}{a function that determines how fast the landscape will regenerate after a
fire event}
}
\description{
Fire Description
}
\examples{

# Use the fire_effects function to modify the habitat using spatial
# fire layers (stored in the landscape object and called "fires")
# and a regeneration function:

fire <- fire_effects(fire_layers = "fires",
                    lag = 5,
                    regeneration_function = function (time) {-time})
}
