% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{as.habitat}
\alias{as.habitat}
\alias{is.habitat}
\alias{print.habitat}
\alias{as.population}
\alias{is.population}
\alias{population}
\alias{population<-}
\alias{pop_patch_name}
\alias{features}
\alias{features<-}
\alias{spatial_patches}
\alias{spatial_patches<-}
\alias{suitability}
\alias{suitability<-}
\alias{[[.habitat}
\alias{carrying_capacity_function}
\title{habitat objects}
\usage{
as.habitat(features, ...)

is.habitat(x)

\method{print}{habitat}(x, ...)

as.population(x, ...)

is.population(x)

population(habitat)

population(habitat) <- value

pop_patch_name(habitat)

features(habitat)

features(habitat) <- value

spatial_patches(habitat)

spatial_patches(habitat, newpatch) <- value

suitability(habitat)

suitability(habitat, updated_habitat_suitability_map) <- value

\method{[[}{habitat}(x, i)

carrying_capacity_function(x, type = c("exp", "logit", "linear", "custom"),
  ...)
}
\arguments{
\item{features}{A named list of landscapes (or seascape) features and parameters used for setting up the habitat for dynamic metapopulation models.}

\item{\dots}{further arguments passed to or from other methods.}

\item{x}{an object to print or test as a habitat object}

\item{i}{index specifying the patches to include in the subset
\code{habitat} object}

\item{type}{model form for converting occurrence to carrying capacity.}

\item{x}{a vector, data.frame, matrix or raster of population(s) numbers for each stage(s) and each patch(s)}

\item{x}{a raster of species habitit suitability (occupancy).}

\item{list}{parameters used to convert a habitat suitability map to carrying capacity.}
}
\value{
an object of class \code{habitat}.
}
\description{
Habitat is an object that contains the spatial distribution of the populations, habitat suitability and carrying capacity for the landscape or seascape. Habitat requires either predefined rasters of population size for each life-history, habitat suitability map (e.g. a species distribution model) and carrying capacity. However, habitat suitability map is the only mandatory raster, population and carrying capacity can be provided as numeric values or functions which manipulate the habitat suitability map rasters to generate population per-cell and/or carrying capacity per-cell.
}
\details{
parameter details for habitat function.
\itemize{
\item{habitat_suitability_map}{ List a named list that contains a \link[raster]{raster} that represents habitat suitability for the species. This need to be probabilitic (between zero and one). Functions that manipulate the landscape will alter this throughout the dynamic metapopulation simulations.}
\item{populations}{ List starting populations for each life-history stage. Either a \link[sp]{SpatialPointsDataFrame} which has the population size for each life history linked to a coordinate within the extent of the habitat_suitability_map. A raster of population per-cell for each life-history stage, or finally a single integer of population size for each life-histroy stage.}
\item{carrying_capacity}{ List either a raster that represent the carrying capacity of adult populations for each cell; a function which manipulates the habitat_suitability_map and converts it to carrying capacity; or finally an integer which presents a maximum carrying capacity for all cells.}
}
}
\examples{
library(raster)
library(dhmpr)
set.seed(42)
xy <- expand.grid(x=seq(145, 150, 0.1), y=seq(-40, -35, 0.1))
Dd <- as.matrix(dist(xy))
w <- exp(-1/nrow(xy) * Dd)
Ww <- chol(w)
xy$z <- t(Ww) \%*\% rnorm(nrow(xy), 0, 0.1)
coordinates(xy) <- ~x+y
r <- rasterize(xy, raster(points2grid(xy)), 'z')
proj4string(r) <- '+init=epsg:4283'
r[] <- scales::rescale(r[],to=c(0,1))

## create a habitat from a list containing a habitat suitability raster and numeric values for population and carrying capacity.
features <- list('habitat_suitability_map'=r,
                       'population'=c(80,20,10),
                       'carrying_capacity'=100)
habs <- as.habitat(features)
                       
## create a habitat from a list containing a habitat suitability raster, a SpatialPointsDataFrame for population and numeric values carrying capacity.
random_populations <- sampleRandom(r, size=50, na.rm=TRUE, sp=TRUE) 
random_populations@data <- as.data.frame(t(rmultinom(50, size = 100, prob = c(0.8,0.2,0.1))))
features <- list('habitat_suitability_map'=r,
                       'population'=random_populations,
                       'carrying_capacity'=100)
                                                
habs <- as.habitat(features)
                               
#create a habitat from a list containing just a species distribution model will estimate populations per-patch.
habs <- as.habitat(list(r))
# print method
print(habitat)

# starting population numbers for each step in the demographic function
population <- as.population(c(80,30,10))
population <- as.population(t(rmultinom(10, size = 100, prob = c(0.8,0.2,0.01))))
population <- as.population(c(80,30,10))
is.population(population)
# get and set the population
population(habitat)
population(habitat) <- population(habitat) * 2
population(habitat)

get and set the features
features(habitat)
# get and set the features
spatial_patches(habitat)

# get and set the features
suitability(habitat)

# habitats can be subsetted to get sub-habitats of patches with double
# braces
habitat
habitat[[1]]
habitat[[1:2]]

}
\author{
Nick Golding & Skip Woolley

Skipton Woolley
}
