% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{habitat}
\alias{habitat}
\alias{is.habitat}
\alias{print.habitat}
\title{Create a habitat object to use in a state object.}
\usage{
habitat(habitat_suitability, carrying_capacity, misc = NULL, ...)

is.habitat(x)

\method{print}{habitat}(x, ...)
}
\arguments{
\item{habitat_suitability}{A raster layer or stack containing habitat
suitability values for all cells in a landscape. Note, using a raster
stack assumes that the user has provided a habitat layer for each
intended timestep in a simulation.}

\item{carrying_capacity}{A raster layer specifying carrying capacity
values for all cells in a landscape.}

\item{misc}{Miscellaneous inputs used to modify the habitat object in a
simulation. Note, this is intended to store objects that are accessed
and used to modify the habitat with a custom developed habitat dynamic
function.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{A habitat object.}
}
\value{
An object of class \code{habitat}
}
\description{
A habitat object is used to store spatially-explicit information on habitat
suitability and the carrying_capacity of a landscape.
}
\details{
A habitat object is a sub-component of a \link[steps]{state} object and
is modified in each timestep of a simulation. During a simulation, a
habitat object tracks changes in habitat suitability or carrying capacity
based on habitat dynamic functions selected or created by the user.
}
\examples{

library(steps)
library(raster)

# Construct the habitat object
hab <- habitat(habitat_suitability = egk_hab, carrying_capacity = egk_k)

# Test if object is of the type 'habitat'
is.habitat(test_habitat)

# Print information about the 'habitat' object
print(test_habitat)
}
