% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat-class.R
\name{build_habitat}
\alias{build_habitat}
\alias{is.habitat}
\alias{print.habitat}
\title{Create a habitat object to use in a state object}
\usage{
build_habitat(habitat_suitability, carrying_capacity, misc = NULL, ...)

is.habitat(x)

\method{print}{habitat}(x, ...)
}
\arguments{
\item{habitat_suitability}{a raster layer or stack containing habitat suitability for each cell}

\item{carrying_capacity}{a raster layer specifying carrying capacity values for each cell}

\item{misc}{miscellaneous inputs used to modify habitat}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object to print or test as an habitat object}
}
\value{
An object of class \code{habitat}
}
\description{
A habitat object is used to store spatially-explicit information on habitat suitability and the carrying_capacity of a landscape.
It is a sub-component of a \link[steps]{state} object and is modified in each timestep of an experiment.
}
\examples{

library(steps)
library(raster)

# Construct a raster object

r <- raster(system.file("external/test.grd", package="raster"))

# Modify raster to contain values between 0 and 1  

hs <- r / cellStats(r, "max")

# Modify raster to contain values for maximum population in each cell

k <- ceiling(r * 0.1)

# Construct the habitat object

test_habitat <- build_habitat(habitat_suitability = hs, carrying_capacity = k)

# Test if object is of the type 'habitat'
  
is.habitat(test_habitat)

# Print information about the 'habitat' object

print(test_habitat)
}
