% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-class.R
\name{habitat_dynamics}
\alias{habitat_dynamics}
\alias{is.habitat_dynamics}
\alias{print.habitat_dynamics}
\title{Modify the habitat in a state object.}
\usage{
habitat_dynamics(...)

is.habitat_dynamics(x)

\method{print}{habitat_dynamics}(x, ...)
}
\arguments{
\item{...}{Functions that operate on a state object to change habitat
at specified timesteps. User may enter a custom function or select a
pre-defined module - see examples.}

\item{x}{an \code{habitat_dynamics} object to print or test.}
}
\value{
An object of class \code{habitat_dynamics}
}
\description{
A \code{habitat dynamics} object is used to modify habitat suitability
(or carrying capacity) of a landscape in space and time - for example,
by fire or intentional habitat modification.
}
\details{
The \code{build_habitat_dynamics} is a sub-component of a
\link[steps]{dynamics} object and is executed in each timestep of a
simulation. Note, some dynamics functions can be executed at
non-regular intervals (i.e. only timesteps explicitly defined by the user).
The \code{build_habitat_dynamics} function is used to construct a
habitat dynamics object consisting of several habitat
dynamics functions and their associated parameters. These functions
specify how the habitat in the state object will be modified
throughout a simulation.
}
\examples{

library(steps)
library(raster)

# Select existing habitat functions to be run on the demography
# in a simulation and specify input parameters: 
fires <- disturbance_fires(habitat_suitability = egk_hab,
                           disturbance_layers = egk_dist,
                           effect_time = 3)

# Construct a habitat dynamics object
hab_dynamics <- habitat_dynamics(fires)

# Test if object is of the type 'habitat dynamics'
is.habitat_dynamics(test_hab_dynamics)

# Print details about the 'habitat_dynamics' object
print(test_hab_dynamics)
}
