% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics-functions.R
\name{habitat_dynamics_functions}
\alias{habitat_dynamics_functions}
\title{Functions to modify the habitat in a landscape object.}
\description{
Pre-defined functions to operate on habitat suitability (and carrying capacity if a function
is used) during a simulation.
}
\seealso{
\itemize{
  \item{\link[steps]{disturbance} to modify the suitability of a landscape with user provided
  spatially-explicit layers}
  \item{\link[steps]{fire_effects}}
  }
}
