% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metapop_n}
\alias{metapop_n}
\title{Simulate a metapopulation system in C++
this function is not complete.}
\usage{
metapop_n(time, dist, area, presence, y = 1, x = 1, e = 1, alpha = 1,
  beta = 1, hanski_dispersal_kernal = "H", locations = NULL, c = 1,
  col_meth = "H")
}
\arguments{
\item{time}{Number of time steps}

\item{dist}{Distances between patches (symetrical matrix)}

\item{area}{Area of patches - This needs to be calculated somehow - using occupancy models?}

\item{presence}{Initial occupancies of patches. Must be presence 1 or absence 0.}

\item{y}{incidence function parameters}

\item{x}{incidence function parameters}

\item{e}{Minimum area of patches}

\item{alpha}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{beta}{double parameter that represents the shape of the dispersal kernel.}

\item{hanski_dispersal_kernal}{char if 'H' uses hanski(1994), if 'S' uses shaw(1995).}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}

\item{c}{double colonisation scale parameter see Ovaskainen 2002.}

\item{col_meth}{char which method to use? 'H' = Hanski 1994, 'M'= Moilanen 2004, 'O'= Ovaskainen 2002. See decription for details.}
}
\description{
Simulate a metapopulation system in C++
this function is not complete.
}

