% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metapop_n}
\alias{metapop_n}
\title{Simulate a metapopulation system in C++
this function is not complete.}
\usage{
metapop_n(time, dist, area, presence, y = 1L, x = 1L, e = 1L, alpha = 1,
  beta = 1, hanski_dispersal_kernal = TRUE, locations = NULL)
}
\arguments{
\item{time}{Number of time steps}

\item{dist}{Distances between patches (symetrical matrix)}

\item{area}{Area of patches - This needs to be calculated somehow - using occupancy models?}

\item{presence}{Initial occupancies of patches. Must be presence 1 or absence 0.}

\item{y}{incidence function parameters}

\item{x}{incidence function parameters}

\item{e}{Minimum area of patches}

\item{alpha}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{beta}{double parameter that represents the shape of the dispersal kernel.}

\item{hanski_dispersal_kernal}{bool if true uses hanski(1994), if false uses shaw(1995).}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}
}
\description{
Simulate a metapopulation system in C++
this function is not complete.
}

