% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation-class.R
\name{metapopulation}
\alias{metapopulation}
\alias{plot.metapopulation}
\title{metapopulation objects}
\usage{
metapopulation(nrep = 10, time = 20, habitat, dispersal, y1 = 1, x1 = 1,
  e1 = 1, locations = NULL, c = 1, coln_fun = "H")

\method{plot}{metapopulation}(x, ...)
}
\arguments{
\item{nrep}{Number of simulations.}

\item{time}{Number of time steps}

\item{habitat}{object}

\item{dispersal}{object see \link[bbgdm]{dispersal} an object that contains information on:
alpha, the decay rate of patch connectivity (dispersion parameter); beta, parameter that controls
 the shape of the dispersal kernel; and disp_fun, which selects the disperal function: 'H' uses hanski(1994)}

\item{y1}{incidence function parameters}

\item{x1}{incidence function parameters}

\item{e1}{Minimum area of patches}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}

\item{c}{double colonisation scale parameter see Ovaskainen 2002.}

\item{coln_fun}{char which colonisation function to use. 'H' = Hanski (1994), 'M'= Moilanen (2004), 'O'= Ovaskainen (2002). See decription for details.}

\item{x}{a metapopulation object}

\item{...}{other plot arguments}
}
\description{
do some metapopulation modelling in R
}
\examples{
habitat <- as.habitat(list(coordinates = data.frame(x=runif( 100, min=-20, max=20),
                                                    y=runif( 100, min=-20, max=20)),
                               area = as.data.frame(exp(-seq(.1,10,length.out = 100))*10),
                               population = as.population(t(rmultinom(100, 
                               size = 100, prob = c(0.8,0.2,0.01)))),
                               features = data.frame(temperature = 10)))
params <- list(alpha=1,beta=1,disp_fun="H")
adult.dispersal <- dispersal(params) 
mp <- metapopulation(nrep=10, time=100, habitat=habitat, dispersal=adult.dispersal,
                  x1 = 200, e1 = 0.00061, y1 = 200)
                  
library(raster)
set.seed(42)
xy <- expand.grid(x=seq(145, 150, 0.1), y=seq(-40, -35, 0.1))
Dd <- as.matrix(dist(xy))
w <- exp(-1/nrow(xy) * Dd)
Ww <- chol(w)
xy$z <- t(Ww) \%*\% rnorm(nrow(xy), 0, 0.1)
coordinates(xy) <- ~x+y
r <- rasterize(xy, raster(points2grid(xy)), 'z')
r2 <- raster(r)
res(r2) <- 0.01
r2 <- resample(r, r2)
proj4string(r2) <- '+init=epsg:4283'                 
habitat <- as.habitat(list(r2,population = as.population(t(rmultinom(1, 
                               size = 100, prob = c(0.8,0.2,0.1))))))
mp1 <- metapopulation(nrep=10, time=100, habitat=habitat, dispersal=adult.dispersal,
                  x1 = 0.42, e1 = 2.61, y1 = 2.2)
                  
habitat <- as.habitat(list(r2))
mp2 <- metapopulation(nrep=10, time=100, habitat=habitat, dispersal=adult.dispersal,
                   x1 = 200, e1 = 0.00061, y1 = 200)
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- metapopulation(nrep=100, time=50, dist, area, presence, x = 0.42, e = 0.061, y = 1.2)
plot(mp)
}
\author{
Skipton Woolley

Skipton Woolley
}
\seealso{
\code{link{metapop_n_cpp}}
}

