% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapopulation-class.R
\name{metapopulation}
\alias{metapopulation}
\alias{plot.metapopulation}
\title{metapopulation objects}
\usage{
metapopulation(nrep = 10, time = 20, habitat, y1 = 1, x1 = 1, e1 = 1,
  alpha = 1, beta = 1, disp_fun = "H", locations = NULL, c = 1,
  coln_fun = "H")

\method{plot}{metapopulation}(x, ...)
}
\arguments{
\item{nrep}{Number of simulations.}

\item{time}{Number of time steps}

\item{habitat}{object}

\item{y1}{incidence function parameters}

\item{x1}{incidence function parameters}

\item{e1}{Minimum area of patches}

\item{alpha}{Exponential decay rate of patch connectivity (dispersion parameter)}

\item{beta}{double parameter that represents the shape of the dispersal kernel.}

\item{disp_fun}{char Which dispersal function to use. 'H' uses hanski(1994), if 'S' uses shaw(1995).}

\item{locations}{NULL or NumericMatrix Longitudes and latitudes of coordinates of the patches}

\item{c}{double colonisation scale parameter see Ovaskainen 2002.}

\item{coln_fun}{char which colonisation function to use. 'H' = Hanski (1994), 'M'= Moilanen (2004), 'O'= Ovaskainen (2002). See decription for details.}

\item{x}{a metapopulation object}

\item{...}{other plot arguments}
}
\description{
do some metapopulation modelling in R
}
\examples{
habitat <- as.habitat(list(coordinates = data.frame(x=runif( 10, min=-10, max=10),
                                                    y=runif( 10, min=-10, max=10)),
                               area = as.data.frame(exp(-seq(.1,10,length.out = 10))*10),
                               population = as.population(t(rmultinom(10, 
                               size = 100, prob = c(0.8,0.2,0.01)))),
                               features = data.frame(temperature = 10)))
mp <- metapopulation(nrep=10, time=100, habitat,
                  x1 = 0.42, e1 = 0.0061, y1 = 1.2)
n <- 50
meta_data <- data.frame(x1=runif( n, min=-10, max=10), x2=runif( n, min=-10, max=10),
area=exp(-seq(.1,10,length.out = n))*10,presence=rbinom(n,1,.8))
area <- meta_data$area
dist <- as.matrix(with(meta_data, dist(cbind(x1, x2))))
presence <- meta_data$presence
mp <- metapopulation(nrep=100, time=50, dist, area, presence, x = 0.42, e = 0.061, y = 1.2)
plot(mp)
}
\author{
Skipton Woolley

Skipton Woolley
}
\seealso{
\code{link{metapop_n_cpp}}
}

