% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_transition_functions-class.R
\name{modified_transition}
\alias{modified_transition}
\title{Spatially-explicit function}
\usage{
modified_transition(survival_layer = NULL, fecundity_layer = NULL)
}
\arguments{
\item{survival_layer}{the name of a spatial layer in the landscape object used to modify survival values.}

\item{fecundity_layer}{the name of a spatial layer in the landscape object used to modify fecundity values.}
}
\value{
An object of class \code{transition_function}
}
\description{
In the built-in \code{modified_transition function}, the values of fecundity and survival
in local cell-based transition matrices are multiplied by values in the named spatial objects
for each cell. The spatial objects can be rasters that are stored in the landscape object.
}
\examples{

test_mod_transition <- modified_transition()
}
