% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-class.R
\name{module}
\alias{as.module}
\alias{module}
\title{module objects}
\usage{
as.module(fun, params, ...)
}
\description{
Module for altering the habitat, they maybe fire spread, management, ect. 
Module sets up internal or custom functions to work with \code{habitat} and \code{simulation} objects.

Module needs to be a function, that gets imported into the simulaion function.
}
\examples{

params = list(habitat=habitat,
              fire_start_location = sample(ncell(r2),10),
              prob = 0.24,
              continue_to_burn_prob = 0.01,
              id=TRUE)
              
fun <- fire_spread                
}

