% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_modification-functions.R
\name{mortality}
\alias{mortality}
\title{Directly affect populations}
\usage{
mortality(mortality_layer, stages = NULL)
}
\arguments{
\item{mortality_layer}{the name of spatial layer(s) in the landscape object with
mortality proportions used to alter the populations for each timestep (number of
layers must match the intended timesteps)}

\item{stages}{which life-stages are modified - default is all}
}
\description{
This function modifies a population by multiplying a spatial layer with populations
in each timestep. The values of the input layers range between 0 and 1 and represent
the proportion reductions in populations (e.g. 0 reduces the population to zero,
whilst 0.5 reduces the population by 50%).
}
\examples{
# Modify populations in all life-stages with fire intensity.

\dontrun{
fire_mortal <- mortality(mortality_layer = "fire", stages = NULL)

ls <- landscape(population = egk_pop,
                suitability = egk_hab,
                carrying_capacity = egk_k,
                "fire" = egk_fire)

pd <- population_dynamics(change = growth(egk_mat), modification = fire_mortal)

simulation(landscape = ls, population_dynamics = pd, habitat_dynamics = NULL, timesteps = 20)
}
}
