% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchify.R
\name{patchify}
\alias{patchify}
\title{patchify
Clump raster cells into patches, with optional neighbourhood distance.}
\usage{
patchify(x, distance, p4s, givedist = TRUE)
}
\arguments{
\item{x}{a binary Raster layer (0 or NA for background, and 1 for areas to be clumped)}

\item{distance}{the neighbourhood distance. Patches that occur within this distance of
one another will be clumped. This should be in the units of the CRS given
in p4s. If this is zero, the function will identify patches defined by 
8-connectivity (i.e. queen's case).}

\item{p4s}{an equal-area projection appropriate for the region. This will be used when 
calculating inter-patch distances. The returned objects will be in the 
original coordinate system.}

\item{givedist}{should the distance matrix be returned? (logical). Distances are in the 
units of p4s, and are shortest cartesian distances between patches.}
}
\description{
patchify
Clump raster cells into patches, with optional neighbourhood distance.
}
\examples{
library(rasterVis)
library(raster)
set.seed(42)
xy <- expand.grid(x=seq(145, 150, 0.1), y=seq(-40, -35, 0.1))
Dd <- as.matrix(dist(xy))
w <- exp(-1/nrow(xy) * Dd)
Ww <- chol(w)
xy$z <- t(Ww) \%*\% rnorm(nrow(xy), 0, 0.1)
coordinates(xy) <- ~x+y
r <- rasterize(xy, raster(points2grid(xy)), 'z')
r2 <- raster(r)
res(r2) <- 0.01
r2 <- resample(r, r2)
proj4string(r2) <- '+init=epsg:4283'
rthr <- r2 > quantile(r2, 0.9)
levelplot(rthr, margin=FALSE, col.regions=0:1, colorkey=FALSE)
foo <- patchify(rthr, distance=1000, '+init=epsg:3577')
spplot(foo$patchrast, col.regions=rainbow(length(foo$patchpoly)))
}
\author{
John Baumgartner
}

