% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{population}
\alias{as.population}
\alias{is.population}
\alias{population}
\title{population objects}
\usage{
as.population(x, ...)

is.population(x)
}
\arguments{
\item{x}{a vector, data.frame or matrix of population(s) numbers for each stage(s) and each patch(s)}
}
\description{
population class
}
\examples{
# starting population numbers for each step in the demographic function
population <- as.population(c(80,30,10))
population <- as.population(t(rmultinom(10, size = 100, prob = c(0.8,0.2,0.01))))
population <- as.population(c(80,30,10))
is.population(population)
}

