% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-class.R
\name{build_population}
\alias{build_population}
\alias{is.population}
\alias{print.population}
\title{Create a population object to use in a state object}
\usage{
build_population(population_raster)

is.population(x)

\method{print}{population}(x, ...)
}
\arguments{
\item{population_raster}{A raster stack with one layer for each life stage}

\item{x}{a population object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{population}
}
\description{
A population object is used to store spatially-explicit information on species populations.
It is a sub-component of a \link[dhmpr]{state} object and is modified in each timestep of an experiment.
}
\examples{

library(dhmpr)
library(raster)

# Construct a raster object

r <- raster(system.file("external/test.grd",package="raster"))

# Create a stack of raster layers to represent each
# life-stage of a population structure (four in this case)

rs <- stack(replicate(4, r * 0.2))

# Construct the population object

test_population <- build_population(rs)

# Test if object is of the type 'population'

is.population(test_population)

# Print information about the 'population' object

print(test_population)
}
