% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_change-functions.R
\name{population_change_functions}
\alias{population_change_functions}
\alias{growth}
\title{How the population changes in a landscape.}
\usage{
growth(transition_matrix, demographic_stochasticity = TRUE,
  global_stochasticity = 0, local_stochasticity = 0)
}
\arguments{
\item{transition_matrix}{A symmetrical age-based (Leslie) or stage-based
population structure matrix.}

\item{demographic_stochasticity}{should demographic stochasticity be used in
population change?}

\item{global_stochasticity, local_stochasticity}{either scalar values or
matrices (with the same dimension as \code{transition_matrix}) giving
variability (in standard deviations) in the transition matrix either for
all populations (\code{global_stochasticity}) or for each population
separately (\code{local_stochasticity})}
}
\description{
Pre-defined functions to define population change (e.g. growth) during a simulation.
}
\examples{

# Use a function to modify the  
# population using life-stage transitions:

test_growth <- growth(egk_mat)
}
