% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_change-functions.R
\name{population_change_functions}
\alias{population_change_functions}
\title{How the population changes in a landscape.}
\description{
Pre-defined or custom functions to define population change during a simulation.
Please see the tutorial vignette titled "Creating custom *steps* functions"
for information on how to write custom functions for use in simulations.
}
\seealso{
\itemize{
  \item{\link[steps]{growth} is a default function for changing populations based on
  transition matrices and functions}
  }
}
