% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_density_dependence-functions.R
\name{population_density_dependence_functions}
\alias{population_density_dependence_functions}
\title{How the population responds to density dependence in a landscape.}
\description{
Pre-defined or custom functions to define population density dependence (e.g. ceiling)
during a simulation. Please see the tutorial vignette titled "Creating custom *steps*
functions" for information on how to write custom functions for use in simulations.
}
\seealso{
\itemize{
  \item{\link[steps]{ceiling_density} to cap populations at carrying capacities}
  }
}
