% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dispersal-functions.R
\name{population_dispersal_functions}
\alias{population_dispersal_functions}
\title{How the population disperses in a landscape.}
\description{
Pre-defined or custom functions to define population dispersal during a
simulation. Each dispersal method uses different computing resources
and may be applicable to different simulation scenarios. Please see the
tutorial vignette titled "Creating custom *steps* functions" for
information on how to write custom functions for use in simulations.
}
\seealso{
\itemize{
  \item{\link[steps]{kernel_dispersal} for kernel-based diffusion dispersal using
  habitat suitability and/or carrying capacity to influence movements}
  \item{\link[steps]{cellular_automata_dispersal} for individual-based movements using
  rule-sets}
  \item{\link[steps]{fast_dispersal} for quick kernel-based diffusion
  dispersal without accounting for spatial heterogeneity}
  }
}
