% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics-class.R
\name{population_dynamics}
\alias{population_dynamics}
\alias{is.population_dynamics}
\alias{print.population_dynamics}
\title{Modify the population in a state object.}
\usage{
population_dynamics(change = simple_growth(), disp = NULL,
  mod = NULL, dens_dep = NULL)

is.population_dynamics(x)

\method{print}{population_dynamics}(x, ...)
}
\arguments{
\item{change}{a \link[steps]{population_dynamics_functions} to define how population growth occurs at each timestep}

\item{disp}{a function to define how the population disperses at each timestep (default is exponential kernel)}

\item{mod}{a function to define any deterministic changes to the population - such as translocation - at each timestep}

\item{dens_dep}{a function to control density dependence effects on the population at each timestep}

\item{x}{a population_dynamic object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{population_dynamics}
}
\description{
A \code{population_dynamics} object is used to modify species populations
in space and time.
}
\details{
A population_dynamics object is a sub-component of a \link[steps]{dynamics}
object and is executed in each timestep of a simulation.  Note, some dynamics
functions can be executed at non-regular intervals (i.e. only timesteps
explicitly defined by the user). The \code{population_dynamics} function is
used to construct a population dynamics object consisting of several population
dynamics functions and their associated parameters. These functions specify how
the population in the state object will be modified throughout a simulation.
}
\examples{

library(steps)
library(raster)

# Construct a population dynamics object - note non-specified parameters
uses default population growth function based on transition matrices
pop_dynamics <- population_dynamics()

# Test if object is of the type 'population dynamics'
is.population_dynamics(test_pop_dynamics)

# Print details about the 'population_dynamics' object 
print(test_pop_dynamics)
}
