% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics-class.R
\name{population_dynamics}
\alias{population_dynamics}
\alias{is.population_dynamics}
\alias{print.population_dynamics}
\title{Define population dynamics.}
\usage{
population_dynamics(change = NULL, dispersal = NULL,
  modification = NULL, density_dependence = NULL,
  dynamics_order = c("change", "dispersal", "modification",
  "density_dependence"))

is.population_dynamics(x)

\method{print}{population_dynamics}(x, ...)
}
\arguments{
\item{change}{\link[steps]{population_change_functions} to define how population
growth occurs at each timestep}

\item{dispersal}{\link[steps]{population_dispersal_functions} to define how the
population disperses at each timestep}

\item{modification}{\link[steps]{population_modification_functions} to define any
deterministic changes to the population - such as translocations or population
control - at each timestep}

\item{density_dependence}{\link[steps]{population_density_dependence_functions}
to control density dependence effects on the population at each timestep}

\item{dynamics_order}{the order in which the population dynamics should be executed
on the landscape object - default is "change" -> "dispersal" -> "modification" -> "density_dependence". 
Note, if population dynamics are reordered, all dynamics must be listed in \code{dynamics_order}.}

\item{x}{a population_dynamics object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class \code{population_dynamics}
}
\description{
A \code{population_dynamics} object is used to describe how populations
change in space and time.
}
\details{
A population_dynamics object is passed to \link[steps]{simulation}
and defines how populations change between timesteps. Note, some dynamics
functions can be executed at non-regular intervals (i.e. only timesteps
explicitly defined by the user). The \code{population_dynamics} function is
used to construct an object with several population dynamics functions and
their associated parameters. These functions specify how the population in
the landscape object will be modified throughout a simulation. The dynamics
can be executed in any order that is specified by the user.
}
\examples{

library(steps)
library(raster)

pop_dynamics <- population_dynamics()

# Test if object is of the type 'population dynamics'
is.population_dynamics(pop_dynamics)

# Print details about the 'population_dynamics' object 
print(pop_dynamics)
}
