% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_modification-functions.R
\name{population_modification_functions}
\alias{population_modification_functions}
\alias{translocation}
\title{How the population is modified in a landscape.}
\usage{
translocation(source_layer, sink_layer, stages = NULL,
  effect_timesteps = NULL)
}
\arguments{
\item{source_layer}{a spatial layer with the locations and number of
individuals to translocate from - note, this layer will only have zero
values if individuals are being introduced from outside the study area}

\item{sink_layer}{a spatial layer with the locations and number of
individuals to translocate to}

\item{stages}{which life-stages are modified - default is all}

\item{effect_timesteps}{which timesteps in a single simulation do the
translocations take place}
}
\description{
Pre-defined functions to define population modification (e.g. translocation) during a simulation.
}
\examples{

# Use the translocation_population_dynamics object to modify the  
# population using translocations:

test_ca_dispersal <- translocation(source_layer = pop_source,
                                       sink_layer = pop_sink,
                                       stages = NULL,
                                       effect_timesteps = 1)
}
