% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_modification-functions.R
\name{population_modification_functions}
\alias{population_modification_functions}
\title{How the population is modified in a landscape.}
\description{
Pre-defined functions to define population modification (e.g. translocation) during a simulation.
}
\seealso{
\itemize{
  \item{\link[steps]{translocation} for specifying explicit spatial and temporal movements
  of populations}
  \item{\link[steps]{mortality} for specifying explicit spatial and temporal changes to populations}
  }
}
