\name{ramasr-package}
\alias{ramasr-package}
\alias{ramasr}
\docType{package}
\title{
\packageTitle{ramasr}
}
\description{
\packageDescription{ramasr}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ramasr}
\packageIndices{ramasr}
~~ An overview of how to use the package, including the most important functions, more details to come soon...~~
}
\author{
\packageAuthor{ramasr}
Maintainer: \packageMaintainer{ramasr}
}
\references{
Akçakaya, H. & Root, W. (2005). RAMAS GIS: Linking spatial data with population viability analysis (version 5.0). Applied Biomathematics, Setauket, New York.
}
\keyword{ package }
