% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_proportion_functions-class.R
\name{set_proportion_dispersing}
\alias{set_proportion_dispersing}
\title{Set proportions of populations dispersing}
\usage{
set_proportion_dispersing(proportions = 1)
}
\arguments{
\item{proportions}{A single value or vector of proportions of individuals in each life stage that disperse
- default is 1. If proportions are specified as a single number, then all life-stages disperse with that
proportion, however, a vector of proportions (equal in length to the number of life-stages) can also be
specified. Note, if a vector of numbers is specified that has fewer elements than life stages, the vector
will be recycled to match its number of elements to life stages (i.e. specifying 'proportions = c(0, 0.5, 1)' with
five life-stages will become 0, 0.5, 1, 0, 0.5).}
}
\value{
An object of class \code{dispersal_proportion_function}
}
\description{
This default function uses a parameter called ???proportions???.
}
\examples{

test_dispersal_function <- set_proportion_dispersing()
}
