% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{simulation}
\alias{is.simulation}
\alias{simulation}
\title{simulation}
\usage{
simulation(x, reps, times, ...)

is.simulation(x)
}
\description{
The over-arching function in dlmpr for simulating a-spatial or spatial demographic population projections through time.
}

