% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{simulation}
\alias{is.simulation}
\alias{simulation}
\title{simulation}
\usage{
simulation(...)

is.simulation(x)
}
\arguments{
\item{dots}{that can contain:
transition a transition object, See \link[dlmpr]{as.transition}.
population a population object, see \link[dlmpr]{as.population}
habitat a habitat object, see \link[dlmpr]{as.habitat}.
dispersal a dispersal object, see \link[dlmpr]{as.dispersal}. 
module a module object, see \link[dlmpr]{as.module}.}
}
\description{
The over-arching function in dlmpr for simulating a-spatial or spatial demographic population projections through time.
}

