% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{simulation}
\alias{simulation}
\alias{is.simulation}
\title{simulation}
\usage{
simulation(x, reps, times, ...)

is.simulation(x)
}
\arguments{
\item{x}{a dynamic object see \link[dhmpr]{dynamic}}

\item{reps}{int number of simulations to run. repetitions}

\item{time_steps}{e.g. years. int number of time steps.}
}
\description{
The over-arching function in dlmpr for simulating a-spatial or spatial demographic population projections through time.
}
\examples{


}
