% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{simulation}
\alias{simulation}
\title{simulation}
\arguments{
\item{transition}{a transition object, See \link[dlmpr]{as.transition}.}

\item{population}{a population object, see \link[dlmpr]{as.population}}

\item{habitat}{a habitat object, see \link[dlmpr]{as.habitat}.}

\item{dispersal}{a dispersal object, see \link[dlmpr]{as.dispersal}.}

\item{module}{a module object, see \link[dlmpr]{as.module}.}
}
\description{
The over-arching function in dlmpr for simulating a-spatial or spatial demographic population projections through time.
}

