% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-class.R
\name{simulation}
\alias{simulation}
\alias{is.simulation}
\title{simulation}
\usage{
simulation(x, reps, ...)

is.simulation(x)
}
\arguments{
\item{x}{an dynamic habitat metapopulation experiment object see \link[dhmpr]{experiment}}

\item{reps}{int number of simulations to run.}

\item{times}{int The number of time steps of the least frequent discrete event. e.g. yearly population growth/management.}
}
\description{
The over-arching function in dhmpr for simulating cell-based spatial demographic meta-population projections through time in a dynamic landscape (or seascape).
}
