% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results-class.R
\name{simulation}
\alias{simulation}
\alias{is.simulation_results}
\alias{print.simulation_results}
\alias{plot.simulation_results}
\title{Run an simulation}
\usage{
simulation(state, dynamics, timesteps, replicates = 1,
  parallel = FALSE)

is.simulation_results(x)

\method{print}{simulation_results}(x, ...)

\method{plot}{simulation_results}(x, object = "population",
  type = "graph", stage = NULL, animate = FALSE,
  timesteps = c(1:9), panels = c(3, 3), ...)
}
\arguments{
\item{state}{a state object - static habitat, population, and demography in a timestep}

\item{dynamics}{a dynamics object - modules that change habitat, population, and demography during a simulation}

\item{timesteps}{number of timesteps used in one simulation or to display when plotting rasters}

\item{replicates}{number simulations to perform}

\item{parallel}{should parallel processors be used for simulations (default is FALSE)}

\item{x}{an simulation_results object}

\item{...}{further arguments passed to or from other methods}

\item{object}{the state object to plot - can be 'population' (default), 'habitat_suitability' or 'carrying_capacity'}

\item{type}{the plot type - 'graph' (default) or 'raster'}

\item{stage}{life-stage to plot - must be specified for 'raster' plot types; default is NULL and all life-stages will be plotted}

\item{animate}{if plotting type 'raster' would you like to animate the rasters as a gif?}

\item{panels}{the number of columns and rows to use when plotting raster timeseries - default is 3 x 3 (e.g. c(3,3) )}
}
\value{
An object of class \code{simulation_results}
}
\description{
A simulation changes state objects based on selected dynamics over a
specified number of timesteps.
}
\examples{

library(steps)
library(raster)
library(future)
plan(multiprocess)

state <- state(population(egk_pop),
               habitat(egk_hab, egk_k),
               demography(egk_mat))

dynamics <- dynamics(habitat_dynamics(),
                     demography_dynamics(),
                     population_dynamics())

results <- simulation(state, dynamics, timesteps = 10, replicates = 5)

# Test if object is of the type 'simulation_results'

is.simulation_results(results)

print(results)

plot(results)
}
