% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results-class.R
\name{simulation}
\alias{simulation}
\alias{is.simulation_results}
\alias{print.simulation_results}
\alias{plot.simulation_results}
\title{Run an simulation}
\usage{
simulation(landscape, population_dynamics, habitat_dynamics = list(),
  timesteps = 3, replicates = 1)

is.simulation_results(x)

\method{print}{simulation_results}(x, ...)

\method{plot}{simulation_results}(x, object = "population",
  type = "graph", stage = NULL, animate = FALSE,
  timesteps = c(1:9), panels = c(3, 3), ...)
}
\arguments{
\item{landscape}{a landscape object representing the initial habitat and
population}

\item{population_dynamics}{a population_dynamics object describing how
population changes over time}

\item{habitat_dynamics}{optional list of functions to modify the landscape at
each timestep}

\item{timesteps}{number of timesteps used in one simulation or to display
when plotting rasters}

\item{replicates}{number simulations to perform}

\item{x}{an simulation_results object}

\item{...}{further arguments passed to or from other methods}

\item{object}{the simulation_results object to plot - can be 'population'
(default), 'suitability' or 'carrying_capacity'}

\item{type}{the plot type - 'graph' (default) or 'raster'}

\item{stage}{life-stage to plot - must be specified for 'raster' plot types;
default is NULL and all life-stages will be plotted}

\item{animate}{if plotting type 'raster' would you like to animate the
rasters as a gif?}

\item{panels}{the number of columns and rows to use when plotting raster
timeseries - default is 3 x 3 (e.g. c(3,3) )}
}
\value{
An object of class \code{simulation_results}
}
\description{
A simulation changes landscape objects based on selected dynamics over a
specified number of timesteps.
}
\examples{

library(steps)
library(raster)
library(future)
plan(multiprocess)

landscape <- landscape(population = egk_pop,
               suitability = egk_hab, 
               carrying_capacity = egk_k)

pop_dynamics <- population_dynamics(change = growth(transition_matrix = egk_mat))

results <- simulation(landscape, pop_dynamics, timesteps = 10, replicates = 5)

# Test if object is of the type 'simulation_results'

is.simulation_results(results)

print(results)

plot(results)
}
