% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPS.R
\name{STEPS}
\alias{STEPS}
\title{Non-linear DSS training function}
\usage{
STEPS(
  data,
  setB.ord.pairs,
  setW.ord.pairs = NULL,
  setR.ord.pairs = NULL,
  n.B,
  lambdaB = 1,
  lambdaW = 1,
  lambdaR = 1,
  rpart_control = NULL,
  rho = 1,
  shrinkage = 0.1,
  stop_terms = list(iter_count = 5, minimal_weight = 10^(-6)),
  test_data = NULL,
  plot.weights = FALSE,
  plot.iteration.trees = FALSE,
  y_plot,
  y_test_plot,
  value_initialize = 0,
  h = 0.5
)
}
\arguments{
\item{data}{data}

\item{setB.ord.pairs}{A table with 2 columns that holds the pairs ordering information.
On the first column are the observations of a better state, and on the second column are the worsen state.
The table has the observations indices in respect to the set O data.}

\item{setW.ord.pairs}{Set W}

\item{setR.ord.pairs}{Set R}

\item{n.B}{number}

\item{lambdaB}{lambs B}

\item{lambdaW}{lambda W}

\item{lambdaR}{lambda R}

\item{rpart_control}{rapart control}

\item{rho}{rho}

\item{shrinkage}{shrinkage}

\item{stop_terms}{stop terms}

\item{test_data}{test}

\item{plot.weights}{binary}

\item{plot.iteration.trees}{binary}

\item{y_plot}{vec}

\item{y_test_plot}{vec test}

\item{value_initialize}{default 0}

\item{h}{space parameter}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Non-linear DSS training function
}
