% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-class.R
\name{as.transition}
\alias{as.transition}
\alias{is.transition}
\alias{plot.transition}
\alias{states}
\alias{summary.transition}
\title{transition objects}
\usage{
as.transition(x, ...)

\method{summary}{transition}(x, ...)

states(transition)

\method{plot}{transition}(x, ...)

is.transition(x)
}
\arguments{
\item{x}{For as.transition, x is a square matrix, that has transition states between population stages.}

\item{...}{other function calls.}

\item{object}{an object of \code{transition} class}
}
\value{
An object of class transition, i.e, resulting from as.transition.
}
\description{
transition is one of the main functions for dhmpr, it helps you construct and process stage based matricies.
Once a transition object is created \code{summary} will return:
 The finite rate of increase ("lambda"),
 the stable stage distribution,
 the reproductive value,
 and the sensitivities and elasticities matrices.
 \code{plot} will return a graph object that plots the transitions between and amoungest each stage of the population matrix.
}
\examples{
mat <- matrix(c(.53,0,.52,0.1,0.77,0,0,0.12,0.9),nrow = 3,ncol = 3,byrow = TRUE)
colnames(mat) <- rownames(mat) <- c('larvae','juvenile','adult') 
trans <- as.transition(mat)

# Add an dispersal object to transition
params <- list(alpha=list('larvae'=2,'juvenile'=0,'adult'=3),
               probability=list('larvae'=0.2,'juvenile'=0,'adult'=0.6))  
all.disperse.different <-  dispersal(params)
trans <- as.transition(mat,all.disperse.different)

summary(tmat)

# get component states
states(trans)
plot(tmat)

is.transition(tmat)

}
\author{
Skipton Woolley

Nick Golding
}

