% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_transition_functions-class.R
\name{modified_transition}
\alias{modified_transition}
\title{Create a growth transition function}
\usage{
modified_transition(transition_matrix, survival_layer = "suitability",
  fecundity_layer = "suitability")
}
\arguments{
\item{transition_matrix}{a symmetrical age-based (Leslie) or stage-based
population structure matrix.}

\item{survival_layer}{a spatial object within the landscape object used to modify survival values.}

\item{fecundity_layer}{a spatial object within the landscape object used to modify fecundity values.}
}
\value{
An object of class \code{transition_function}
}
\description{
A growth transition function defines how spatial objects influence survival and fecundity. 
Functions that use elements of the landscape object to modify survival and fecundity are built-in for the user to select, however, a user may also provide a custom written growth function.
}
\examples{

test_transition_function <- modified_transition(egk_mat)
}
