% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_transition_functions-class.R
\name{transition_function}
\alias{transition_function}
\title{Create a growth transition function}
\description{
A growth transition function defines how spatial objects or custom functions influence
survival and fecundity. Two built-in functions are provided for the user to select, however,
a user may also provide custom written functions to modify survival and fecundity throughout
a simulation. Please see the tutorial vignette titled "Creating custom *steps* functions"
for information on how to write custom functions for use in simulations.
}
\seealso{
\itemize{
  \item{\link[steps]{modified_transition} to use rasters to modify survival and fecundity}
  \item{\link[steps]{competition_density} to use relationship to carrying capacity to modify
  survival and fecundity}
  }
}
